refactor(i18n): reduce i18n.ts to thin runtime and split locale data
Move the 13 non-English locale objects (es, es_ES, fr_FR, fr_CA, en_CA, it_IT,
de_DE, zh_CN, zh_TW, ko_KR, ja_JP, pt_BR, ru_RU) into src/ui/i18n.locales/<code>.ts,
each still a nested ": typeof en" object that imports its content layers and
gameStrings variant from ../i18n.en. i18n.ts is now a thin runtime: it assembles
translations from the per-locale modules, keeps t/tOptional/hasTranslation/the
formatters/language getters-setters/supportedLanguages, and re-exports every
locale object plus the types. All public exports preserved in name and signature.

Behavior-preserving: resolved 14-locale table is byte-identical.
