+++ src/ui/i18n.locales/de_DE.ts
@@ -1,0 +1,169 @@
+import {
+  en,
+  shellStrings,
+  hudStrings,
+  abilityStrings,
+  questStrings,
+  itemStrings,
+  classAbilityNames,
+  itemNames,
+  mergeStrings,
+  mergeEntities,
+  mergeExtra,
+  gameStringsDeDE,
+} from '../i18n.en';
+import { worldEntityText as worldNames } from '../world_entity_i18n';
+
+export const de_DE: typeof en = {
+  meta: { builtOn: "Erstellt am {date}" },
+  realmTypes: { normal: "Normal", pvp: "PvP", rp: "RP", rpPvp: "RP-PvP" },
+  game: gameStringsDeDE,
+  nav: {
+    home: "Startseite",
+    play: "Spielen",
+    stats: "Statistiken",
+    about: "Über uns",
+    highscores: "Bestenlisten",
+    wiki: "Wiki",
+    news: "Neuigkeiten",
+    download: "Herunterladen",
+    loginRegister: "Einloggen/Registrieren",
+    donate: "Spenden",
+  },
+  stats: {
+    title: "Realmstatus",
+    accountsCreated: "Erstellte Accounts",
+    playersOnline: "Spieler online",
+    realmName: "Realmname",
+  },
+  footer: {
+    copyright: "2026 World of ClaudeCraft",
+    githubLink: "https://github.com/levy-street/world-of-claudecraft",
+    githubLabel: "Open-Source-Projekt",
+    terms: "Nutzungsbedingungen",
+    privacy: "Datenschutzerklärung",
+    discordLabel: "Tritt dem Discord bei",
+  },
+  highscores: {
+    title: "Bestenliste",
+    desc: "Verfolge die größten Champions des Realms und vergleiche deinen Fortschritt.",
+  },
+  wiki: {
+    title: "Spiel-Wiki und Guides",
+    desc: "Entdecke die Geheimnisse des Realms, Klassenguides und Strategien.",
+  },
+  news: {
+    title: "Neuigkeiten und Updates",
+    desc: "Lies die neuesten Patchnotes, Events und Community-Updates.",
+  },
+  download: {
+    title: "Desktop-Launcher herunterladen",
+    desc: "Hol dir den eigenständigen Launcher für optimierte Leistung und Vollbildspiel.",
+  },
+  comingSoon: {
+    placeholder: "Demnächst...",
+    featureComingSoon: "Diese Funktion ist demnächst auf dem Realm verfügbar.",
+  },
+  mode: {
+    onlineTitle: "Online spielen",
+    onlineDesc: "Logge dich im Realm ein. Deine Charaktere befinden sich auf dem Server und du teilst dir die Welt mit allen anderen Spielern online.",
+    onlineAria: "Online spielen: Logge dich im persistenten, geteilten Realm ein",
+    offlineTitle: "Offline spielen",
+    offlineDesc: "Sofortige Einzelspielerwelt in deinem Browser. Es wird nichts gespeichert: perfekt für einen schnellen Kampf oder zum Testen.",
+    offlineAria: "Offline spielen: Starte eine sofortige lokale Einzelspielersitzung",
+    tipTitle: "TIPP:",
+    tipText: "Deaktiviere für das flüssigste Spielerlebnis Adblocker-Erweiterungen auf dieser Website. Community-Berichte zeigen, dass einige Blocker zu Verzögerungen führen können.",
+  },
+  auth: {
+    enterRealm: "Realm betreten",
+    username: "Benutzername",
+    usernameError: "Bitte gib deinen Benutzernamen ein.",
+    usernamePlaceholder: "Benutzername eingeben",
+    password: "Passwort",
+    passwordError: "Bitte gib dein Passwort ein.",
+    passwordPlaceholder: "Passwort eingeben",
+    showPassword: "Passwort anzeigen",
+    hidePassword: "Passwort ausblenden",
+    logIn: "Einloggen",
+    createAccount: "Account erstellen",
+    back: "Zurück",
+    realmList: "Realmliste",
+    loadingRealms: "Realms werden geladen...",
+    changeRealm: "Realm wechseln",
+    characters: "Charaktere:",
+    createCharacter: "Charakter erstellen",
+    characterName: "Charaktername",
+    characterNamePlaceholder: "Charaktername",
+    enterWorld: "Welt betreten",
+    offlineCharacter: "Offline-Charakter",
+    create: "Erstellen",
+  },
+  classes: {
+    warrior: "Krieger",
+    paladin: "Paladin",
+    hunter: "Jäger",
+    rogue: "Schurke",
+    priest: "Priester",
+    shaman: "Schamane",
+    mage: "Magier",
+    warlock: "Hexenmeister",
+    druid: "Druide",
+    warriorAria: "Krieger-Klasse",
+    paladinAria: "Paladin-Klasse",
+    hunterAria: "Jäger-Klasse",
+    rogueAria: "Schurke-Klasse",
+    priestAria: "Priester-Klasse",
+    shamanAria: "Schamane-Klasse",
+    mageAria: "Magier-Klasse",
+    warlockAria: "Hexenmeister-Klasse",
+    druidAria: "Druide-Klasse",
+  },
+  controls: {
+    title: "Steuerungsanleitung",
+    movement: "Bewegung",
+    moveTurn: "Bewegen/Drehen",
+    strafe: "Seitwärts bewegen Links/Rechts",
+    jump: "Springen",
+    autorun: "Automatisches Laufen umschalten",
+    combat: "Kampf und Interaktion",
+    target: "Feind anvisieren",
+    spells: "Zauber wirken",
+    interact: "Interagieren/Plündern",
+    nameplates: "Namensplaketten umschalten",
+    camera: "Kamera und Maus",
+    rightDrag: "Rechts ziehen",
+    leftDrag: "Links ziehen",
+    mouseWheel: "Mausrad",
+    mouselook: "Mausblick",
+    orbit: "Kamera umkreisen",
+    zoom: "Zoom",
+    interfaces: "Benutzeroberflächen",
+    charPane: "Charakterfenster",
+    spellbook: "Zauberbuch",
+    questLog: "Questlog",
+    worldMap: "Weltkarte",
+    bags: "Tascheninventar",
+    emoteWheel: "Emote-Rad gedrückt halten",
+    friends: "Freunde und Gilde",
+    chat: "Chat öffnen",
+  },
+  ...shellStrings.de_DE,
+  ...hudStrings.de_DE,
+  ...abilityStrings.de_DE,
+  ...questStrings.de_DE,
+  ...itemStrings.de_DE,
+  ...classAbilityNames.de_DE,
+  ...itemNames.de_DE,
+  ...worldNames.de_DE,
+  ...mergeStrings.de_DE,
+  entities: {
+    ...itemNames.de_DE.entities,
+    ...worldNames.de_DE.entities,
+    abilities: { ...itemNames.de_DE.entities.abilities, ...mergeExtra.de_DE.abilities },
+    items: { ...itemNames.de_DE.entities.items, ...mergeEntities.de_DE.items, ...mergeExtra.de_DE.items },
+    mobs: { ...worldNames.de_DE.entities.mobs, ...mergeEntities.de_DE.mobs, ...mergeExtra.de_DE.mobs },
+    npcs: { ...worldNames.de_DE.entities.npcs, ...mergeExtra.de_DE.npcs },
+    quests: { ...worldNames.de_DE.entities.quests, ...mergeEntities.de_DE.quests, ...mergeExtra.de_DE.quests },
+    dungeons: { ...worldNames.de_DE.entities.dungeons, ...mergeExtra.de_DE.dungeons },
+  },
+};
