+++ src/ui/i18n.locales/en_CA.ts
@@ -1,0 +1,169 @@
+import {
+  en,
+  shellStrings,
+  hudStrings,
+  abilityStrings,
+  questStrings,
+  itemStrings,
+  classAbilityNames,
+  itemNames,
+  mergeStrings,
+  mergeEntities,
+  mergeExtra,
+  gameStringsEnCA,
+} from '../i18n.en';
+import { worldEntityText as worldNames } from '../world_entity_i18n';
+
+export const en_CA: typeof en = {
+  meta: { builtOn: "Built {date}" },
+  realmTypes: { normal: "Normal", pvp: "PvP", rp: "RP", rpPvp: "RP-PvP" },
+  game: gameStringsEnCA,
+  nav: {
+    home: "Home",
+    play: "Play",
+    stats: "Statistics",
+    about: "About",
+    highscores: "High Scores",
+    wiki: "Wiki",
+    news: "News",
+    download: "Download",
+    loginRegister: "Login/Register",
+    donate: "Donate",
+  },
+  stats: {
+    title: "Realm Status",
+    accountsCreated: "Accounts Created",
+    playersOnline: "Players Online",
+    realmName: "Realm Name",
+  },
+  footer: {
+    copyright: "2026 World of ClaudeCraft",
+    githubLink: "https://github.com/levy-street/world-of-claudecraft",
+    githubLabel: "Open Source Project",
+    terms: "Terms of Service",
+    privacy: "Privacy Policy",
+    discordLabel: "Join the Discord",
+  },
+  highscores: {
+    title: "High Scores Leaderboard",
+    desc: "Track the realm's greatest champions and compare your progress.",
+  },
+  wiki: {
+    title: "Game Wiki & Guide",
+    desc: "Discover the secrets of the realm, class guides, and strategies.",
+  },
+  news: {
+    title: "News & Updates",
+    desc: "Read the latest patch notes, events, and community updates.",
+  },
+  download: {
+    title: "Download Desktop Launcher",
+    desc: "Get the standalone launcher for optimized performance and full-screen play.",
+  },
+  comingSoon: {
+    placeholder: "Coming Soon...",
+    featureComingSoon: "This feature is coming soon to the realm.",
+  },
+  mode: {
+    onlineTitle: "Play Online",
+    onlineDesc: "Log in to the realm. Your characters live on the server and you share the world with everyone else who's on.",
+    onlineAria: "Play Online: log in to the persistent shared realm",
+    offlineTitle: "Play Offline",
+    offlineDesc: "Instant single-player world in your browser. Nothing is saved: perfect for a quick brawl or testing.",
+    offlineAria: "Play Offline: start an instant local single-player session",
+    tipTitle: "TIP:",
+    tipText: "For the smoothest experience, turn off ad blocker extensions on this site. Community reports found some blockers can cause lag.",
+  },
+  auth: {
+    enterRealm: "Enter the Realm",
+    username: "Username",
+    usernameError: "Please enter your username.",
+    usernamePlaceholder: "Enter username",
+    password: "Password",
+    passwordError: "Please enter your password.",
+    passwordPlaceholder: "Enter password",
+    showPassword: "Show password",
+    hidePassword: "Hide password",
+    logIn: "Log In",
+    createAccount: "Create Account",
+    back: "Back",
+    realmList: "Realm List",
+    loadingRealms: "Loading realms...",
+    changeRealm: "Change Realm",
+    characters: "Characters:",
+    createCharacter: "Create Character",
+    characterName: "Character Name",
+    characterNamePlaceholder: "Character name",
+    enterWorld: "Enter World",
+    offlineCharacter: "Offline Character",
+    create: "Create",
+  },
+  classes: {
+    warrior: "Warrior",
+    paladin: "Paladin",
+    hunter: "Hunter",
+    rogue: "Rogue",
+    priest: "Priest",
+    shaman: "Shaman",
+    mage: "Mage",
+    warlock: "Warlock",
+    druid: "Druid",
+    warriorAria: "Warrior class",
+    paladinAria: "Paladin class",
+    hunterAria: "Hunter class",
+    rogueAria: "Rogue class",
+    priestAria: "Priest class",
+    shamanAria: "Shaman class",
+    mageAria: "Mage class",
+    warlockAria: "Warlock class",
+    druidAria: "Druid class",
+  },
+  controls: {
+    title: "Controls Guide",
+    movement: "Movement",
+    moveTurn: "Move / Turn",
+    strafe: "Strafe Left/Right",
+    jump: "Jump",
+    autorun: "Toggle Autorun",
+    combat: "Combat & Interaction",
+    target: "Target Enemy",
+    spells: "Cast Spells",
+    interact: "Interact / Loot",
+    nameplates: "Toggle Nameplates",
+    camera: "Camera & Mouse",
+    rightDrag: "Right-Drag",
+    leftDrag: "Left-Drag",
+    mouseWheel: "Mouse Wheel",
+    mouselook: "Mouselook",
+    orbit: "Orbit Camera",
+    zoom: "Zoom",
+    interfaces: "Interfaces",
+    charPane: "Character Pane",
+    spellbook: "Spellbook",
+    questLog: "Quest Log",
+    worldMap: "World Map",
+    bags: "Bags Inventory",
+    emoteWheel: "Hold Emote Wheel",
+    friends: "Friends & Guild",
+    chat: "Open Chat",
+  },
+  ...shellStrings.en_CA,
+  ...hudStrings.en_CA,
+  ...abilityStrings.en_CA,
+  ...questStrings.en_CA,
+  ...itemStrings.en_CA,
+  ...classAbilityNames.en_CA,
+  ...itemNames.en_CA,
+  ...worldNames.en_CA,
+  ...mergeStrings.en_CA,
+  entities: {
+    ...itemNames.en_CA.entities,
+    ...worldNames.en_CA.entities,
+    abilities: { ...itemNames.en_CA.entities.abilities, ...mergeExtra.en_CA.abilities },
+    items: { ...itemNames.en_CA.entities.items, ...mergeEntities.en_CA.items, ...mergeExtra.en_CA.items },
+    mobs: { ...worldNames.en_CA.entities.mobs, ...mergeEntities.en_CA.mobs, ...mergeExtra.en_CA.mobs },
+    npcs: { ...worldNames.en_CA.entities.npcs, ...mergeExtra.en_CA.npcs },
+    quests: { ...worldNames.en_CA.entities.quests, ...mergeEntities.en_CA.quests, ...mergeExtra.en_CA.quests },
+    dungeons: { ...worldNames.en_CA.entities.dungeons, ...mergeExtra.en_CA.dungeons },
+  },
+};
