+++ src/ui/i18n.locales/es.ts
@@ -1,0 +1,169 @@
+import {
+  en,
+  shellStrings,
+  hudStrings,
+  abilityStrings,
+  questStrings,
+  itemStrings,
+  classAbilityNames,
+  itemNames,
+  mergeStrings,
+  mergeEntities,
+  mergeExtra,
+  gameStringsEs,
+} from '../i18n.en';
+import { worldEntityText as worldNames } from '../world_entity_i18n';
+
+export const es: typeof en = {
+  meta: { builtOn: "Compilado el {date}" },
+  realmTypes: { normal: "Normal", pvp: "JcJ", rp: "JdR", rpPvp: "JdR-JcJ" },
+  game: gameStringsEs,
+  nav: {
+    home: "Inicio",
+    play: "Jugar",
+    stats: "Estadísticas",
+    about: "Acerca de",
+    highscores: "Clasificaciones",
+    wiki: "Wiki",
+    news: "Noticias",
+    download: "Descargar",
+    loginRegister: "Iniciar Sesión/Registrarse",
+    donate: "Donar",
+  },
+  stats: {
+    title: "Estado del Reino",
+    accountsCreated: "Cuentas Creadas",
+    playersOnline: "Jugadores en Línea",
+    realmName: "Nombre del Reino",
+  },
+  footer: {
+    copyright: "2026 World of ClaudeCraft",
+    githubLink: "https://github.com/levy-street/world-of-claudecraft",
+    githubLabel: "Proyecto de Código Abierto",
+    terms: "Términos de Servicio",
+    privacy: "Política de Privacidad",
+    discordLabel: "Únete al Discord",
+  },
+  highscores: {
+    title: "Clasificaciones de Puntuación",
+    desc: "Sigue a los campeones más grandes del reino y compara tu progreso.",
+  },
+  wiki: {
+    title: "Wiki y Guía del Juego",
+    desc: "Descubre los secretos del reino, guías de clase y estrategias.",
+  },
+  news: {
+    title: "Noticias y Actualizaciones",
+    desc: "Lee las últimas notas del parche, eventos y noticias de la comunidad.",
+  },
+  download: {
+    title: "Descargar Lanzador de Escritorio",
+    desc: "Consigue el lanzador independiente para un rendimiento optimizado y juego a pantalla completa.",
+  },
+  comingSoon: {
+    placeholder: "Próximamente...",
+    featureComingSoon: "Esta característica llegará pronto al reino.",
+  },
+  mode: {
+    onlineTitle: "Jugar en Línea",
+    onlineDesc: "Inicia sesión en el reino. Tus personajes viven en el servidor y compartes el mundo con todos los demás que estén conectados.",
+    onlineAria: "Jugar en Línea: inicia sesión en el reino compartido persistente",
+    offlineTitle: "Jugar en Solitario",
+    offlineDesc: "Mundo instantáneo en solitario en tu navegador. Nada se guarda: perfecto para una pelea rápida o pruebas.",
+    offlineAria: "Jugar en Solitario: inicia una sesión local instantánea de un jugador",
+    tipTitle: "CONSEJO:",
+    tipText: "Para la mejor experiencia, desactiva las extensiones de bloqueo de publicidad en este sitio. Reportes de la comunidad encontraron que algunos bloqueadores pueden causar retrasos.",
+  },
+  auth: {
+    enterRealm: "Entrar al Reino",
+    username: "Usuario",
+    usernameError: "Por favor, introduce tu nombre de usuario.",
+    usernamePlaceholder: "Introduce tu usuario",
+    password: "Contraseña",
+    passwordError: "Por favor, introduce tu contraseña.",
+    passwordPlaceholder: "Introduce tu contraseña",
+    showPassword: "Mostrar contraseña",
+    hidePassword: "Ocultar contraseña",
+    logIn: "Iniciar Sesión",
+    createAccount: "Crear Cuenta",
+    back: "Atrás",
+    realmList: "Lista de Reinos",
+    loadingRealms: "Cargando reinos...",
+    changeRealm: "Cambiar de Reino",
+    characters: "Personajes:",
+    createCharacter: "Crear Personaje",
+    characterName: "Nombre del Personaje",
+    characterNamePlaceholder: "Nombre del personaje",
+    enterWorld: "Entrar al Mundo",
+    offlineCharacter: "Personaje en Solitario",
+    create: "Crear",
+  },
+  classes: {
+    warrior: "Guerrero",
+    paladin: "Paladín",
+    hunter: "Cazador",
+    rogue: "Pícaro",
+    priest: "Sacerdote",
+    shaman: "Chamán",
+    mage: "Mago",
+    warlock: "Brujo",
+    druid: "Druida",
+    warriorAria: "Clase Guerrero",
+    paladinAria: "Clase Paladín",
+    hunterAria: "Clase Cazador",
+    rogueAria: "Clase Pícaro",
+    priestAria: "Clase Sacerdote",
+    shamanAria: "Clase Chamán",
+    mageAria: "Clase Mago",
+    warlockAria: "Clase Brujo",
+    druidAria: "Clase Druida",
+  },
+  controls: {
+    title: "Guía de Controles",
+    movement: "Movimiento",
+    moveTurn: "Moverse / Girar",
+    strafe: "Desplazarse Izquierda/Derecha",
+    jump: "Saltar",
+    autorun: "Correr Automáticamente",
+    combat: "Combate e Interacción",
+    target: "Marcar Enemigo",
+    spells: "Lanzar Hechizos",
+    interact: "Interactuar / Despojar",
+    nameplates: "Mostrar Nombres",
+    camera: "Cámara y Ratón",
+    rightDrag: "Arrastrar Clic Derecho",
+    leftDrag: "Arrastrar Clic Izquierdo",
+    mouseWheel: "Rueda del Ratón",
+    mouselook: "Mirar con Ratón",
+    orbit: "Rotar Cámara",
+    zoom: "Zoom",
+    interfaces: "Interfaces",
+    charPane: "Panel de Personaje",
+    spellbook: "Libro de Hechizos",
+    questLog: "Diario de Misiones",
+    worldMap: "Mapa del Mundo",
+    bags: "Inventario de Bolsas",
+    emoteWheel: "Mantener rueda de gestos",
+    friends: "Amigos y Hermandad",
+    chat: "Abrir Chat",
+  },
+  ...shellStrings.es,
+  ...hudStrings.es,
+  ...abilityStrings.es,
+  ...questStrings.es,
+  ...itemStrings.es,
+  ...classAbilityNames.es,
+  ...itemNames.es,
+  ...worldNames.es,
+  ...mergeStrings.es,
+  entities: {
+    ...itemNames.es.entities,
+    ...worldNames.es.entities,
+    abilities: { ...itemNames.es.entities.abilities, ...mergeExtra.es.abilities },
+    items: { ...itemNames.es.entities.items, ...mergeEntities.es.items, ...mergeExtra.es.items },
+    mobs: { ...worldNames.es.entities.mobs, ...mergeEntities.es.mobs, ...mergeExtra.es.mobs },
+    npcs: { ...worldNames.es.entities.npcs, ...mergeExtra.es.npcs },
+    quests: { ...worldNames.es.entities.quests, ...mergeEntities.es.quests, ...mergeExtra.es.quests },
+    dungeons: { ...worldNames.es.entities.dungeons, ...mergeExtra.es.dungeons },
+  },
+};
