+++ src/ui/i18n.locales/es_ES.ts
@@ -1,0 +1,169 @@
+import {
+  en,
+  shellStrings,
+  hudStrings,
+  abilityStrings,
+  questStrings,
+  itemStrings,
+  classAbilityNames,
+  itemNames,
+  mergeStrings,
+  mergeEntities,
+  mergeExtra,
+  gameStringsEsES,
+} from '../i18n.en';
+import { worldEntityText as worldNames } from '../world_entity_i18n';
+
+export const es_ES: typeof en = {
+  meta: { builtOn: "Compilado el {date}" },
+  realmTypes: { normal: "Normal", pvp: "JcJ", rp: "JdR", rpPvp: "JdR-JcJ" },
+  game: gameStringsEsES,
+  nav: {
+    home: "Inicio",
+    play: "Jugar",
+    stats: "Estadísticas",
+    about: "Acerca de",
+    highscores: "Clasificaciones",
+    wiki: "Wiki",
+    news: "Noticias",
+    download: "Descargar",
+    loginRegister: "Iniciar sesión/Registrarse",
+    donate: "Donar",
+  },
+  stats: {
+    title: "Estado del Reino",
+    accountsCreated: "Cuentas creadas",
+    playersOnline: "Jugadores en línea",
+    realmName: "Nombre del reino",
+  },
+  footer: {
+    copyright: "2026 World of ClaudeCraft",
+    githubLink: "https://github.com/levy-street/world-of-claudecraft",
+    githubLabel: "Proyecto de código abierto",
+    terms: "Términos de servicio",
+    privacy: "Política de privacidad",
+    discordLabel: "Únete al Discord",
+  },
+  highscores: {
+    title: "Tabla de clasificaciones",
+    desc: "Sigue a los campeones más grandes del reino y compara tu progreso.",
+  },
+  wiki: {
+    title: "Wiki y guía del juego",
+    desc: "Descubre los secretos del reino, guías de clase y estrategias.",
+  },
+  news: {
+    title: "Noticias y actualizaciones",
+    desc: "Lee las últimas notas del parche, eventos y noticias de la comunidad.",
+  },
+  download: {
+    title: "Descargar lanzador de escritorio",
+    desc: "Consigue el lanzador independiente para un rendimiento optimizado y juego a pantalla completa.",
+  },
+  comingSoon: {
+    placeholder: "Próximamente...",
+    featureComingSoon: "Esta característica llegará pronto al reino.",
+  },
+  mode: {
+    onlineTitle: "Jugar en línea",
+    onlineDesc: "Inicia sesión en el reino. Tus personajes viven en el servidor y compartes el mundo con todos los demás que estén conectados.",
+    onlineAria: "Jugar en línea: inicia sesión en el reino compartido persistente",
+    offlineTitle: "Jugar en solitario",
+    offlineDesc: "Mundo instantáneo en solitario en tu navegador. Nada se guarda: perfecto para una pelea rápida o pruebas.",
+    offlineAria: "Jugar en solitario: inicia una sesión local instantánea de un jugador",
+    tipTitle: "CONSEJO:",
+    tipText: "Para la mejor experiencia, desactiva las extensiones de bloqueo de publicidad en este sitio. Reportes de la comunidad encontraron que algunos bloqueadores pueden causar retrasos.",
+  },
+  auth: {
+    enterRealm: "Entrar al reino",
+    username: "Usuario",
+    usernameError: "Por favor, introduce tu nombre de usuario.",
+    usernamePlaceholder: "Introduce tu usuario",
+    password: "Contraseña",
+    passwordError: "Por favor, introduce tu contraseña.",
+    passwordPlaceholder: "Introduce tu contraseña",
+    showPassword: "Mostrar contraseña",
+    hidePassword: "Ocultar contraseña",
+    logIn: "Iniciar sesión",
+    createAccount: "Crear cuenta",
+    back: "Atrás",
+    realmList: "Lista de reinos",
+    loadingRealms: "Cargando reinos...",
+    changeRealm: "Cambiar de reino",
+    characters: "Personajes:",
+    createCharacter: "Crear personaje",
+    characterName: "Nombre del personaje",
+    characterNamePlaceholder: "Nombre del personaje",
+    enterWorld: "Entrar al mundo",
+    offlineCharacter: "Personaje en solitario",
+    create: "Crear",
+  },
+  classes: {
+    warrior: "Guerrero",
+    paladin: "Paladín",
+    hunter: "Cazador",
+    rogue: "Pícaro",
+    priest: "Sacerdote",
+    shaman: "Chamán",
+    mage: "Mago",
+    warlock: "Brujo",
+    druid: "Druida",
+    warriorAria: "Clase Guerrero",
+    paladinAria: "Clase Paladín",
+    hunterAria: "Clase Cazador",
+    rogueAria: "Clase Pícaro",
+    priestAria: "Clase Sacerdote",
+    shamanAria: "Clase Chamán",
+    mageAria: "Clase Mago",
+    warlockAria: "Clase Brujo",
+    druidAria: "Clase Druida",
+  },
+  controls: {
+    title: "Guía de controles",
+    movement: "Movimiento",
+    moveTurn: "Moverse/Girar",
+    strafe: "Desplazarse Izquierda/Derecha",
+    jump: "Saltar",
+    autorun: "Correr automáticamente",
+    combat: "Combate e interacción",
+    target: "Marcar enemigo",
+    spells: "Lanzar hechizos",
+    interact: "Interactuar/Despojar",
+    nameplates: "Mostrar nombres",
+    camera: "Cámara y ratón",
+    rightDrag: "Arrastrar clic derecho",
+    leftDrag: "Arrastrar clic izquierdo",
+    mouseWheel: "Rueda del ratón",
+    mouselook: "Mirar con ratón",
+    orbit: "Rotar cámara",
+    zoom: "Zoom",
+    interfaces: "Interfaces",
+    charPane: "Panel de personaje",
+    spellbook: "Libro de hechizos",
+    questLog: "Diario de misiones",
+    worldMap: "Mapa del mundo",
+    bags: "Inventario de bolsas",
+    emoteWheel: "Mantener rueda de gestos",
+    friends: "Amigos y hermandad",
+    chat: "Abrir chat",
+  },
+  ...shellStrings.es_ES,
+  ...hudStrings.es_ES,
+  ...abilityStrings.es_ES,
+  ...questStrings.es_ES,
+  ...itemStrings.es_ES,
+  ...classAbilityNames.es_ES,
+  ...itemNames.es_ES,
+  ...worldNames.es_ES,
+  ...mergeStrings.es_ES,
+  entities: {
+    ...itemNames.es_ES.entities,
+    ...worldNames.es_ES.entities,
+    abilities: { ...itemNames.es_ES.entities.abilities, ...mergeExtra.es_ES.abilities },
+    items: { ...itemNames.es_ES.entities.items, ...mergeEntities.es_ES.items, ...mergeExtra.es_ES.items },
+    mobs: { ...worldNames.es_ES.entities.mobs, ...mergeEntities.es_ES.mobs, ...mergeExtra.es_ES.mobs },
+    npcs: { ...worldNames.es_ES.entities.npcs, ...mergeExtra.es_ES.npcs },
+    quests: { ...worldNames.es_ES.entities.quests, ...mergeEntities.es_ES.quests, ...mergeExtra.es_ES.quests },
+    dungeons: { ...worldNames.es_ES.entities.dungeons, ...mergeExtra.es_ES.dungeons },
+  },
+};
