+++ src/ui/i18n.locales/fr_CA.ts
@@ -1,0 +1,169 @@
+import {
+  en,
+  shellStrings,
+  hudStrings,
+  abilityStrings,
+  questStrings,
+  itemStrings,
+  classAbilityNames,
+  itemNames,
+  mergeStrings,
+  mergeEntities,
+  mergeExtra,
+  gameStringsFrCA,
+} from '../i18n.en';
+import { worldEntityText as worldNames } from '../world_entity_i18n';
+
+export const fr_CA: typeof en = {
+  meta: { builtOn: "Compilé le {date}" },
+  realmTypes: { normal: "Normal", pvp: "JcJ", rp: "JdR", rpPvp: "JdR-JcJ" },
+  game: gameStringsFrCA,
+  nav: {
+    home: "Accueil",
+    play: "Jouer",
+    stats: "Statistiques",
+    about: "À propos",
+    highscores: "Classements",
+    wiki: "Wiki",
+    news: "Actualités",
+    download: "Télécharger",
+    loginRegister: "Se connecter/S'enregistrer",
+    donate: "Faire un don",
+  },
+  stats: {
+    title: "Statut du royaume",
+    accountsCreated: "Comptes créés",
+    playersOnline: "Joueurs en ligne",
+    realmName: "Nom du royaume",
+  },
+  footer: {
+    copyright: "2026 World of ClaudeCraft",
+    githubLink: "https://github.com/levy-street/world-of-claudecraft",
+    githubLabel: "Projet open source",
+    terms: "Conditions d'utilisation",
+    privacy: "Politique de confidentialité",
+    discordLabel: "Rejoindre le Discord",
+  },
+  highscores: {
+    title: "Tableau des classements",
+    desc: "Suivez les plus grands champions du royaume et comparez votre progression.",
+  },
+  wiki: {
+    title: "Wiki et guide du jeu",
+    desc: "Découvrez les secrets du royaume, les guides de classe et les stratégies.",
+  },
+  news: {
+    title: "Actualités et mises à jour",
+    desc: "Lisez les dernières notes de mise à jour, les événements et les nouvelles de la communauté.",
+  },
+  download: {
+    title: "Télécharger le lanceur de bureau",
+    desc: "Obtenez le lanceur autonome pour des performances optimisées et un jeu en plein écran.",
+  },
+  comingSoon: {
+    placeholder: "Bientôt disponible...",
+    featureComingSoon: "Cette fonctionnalité sera bientôt disponible sur le royaume.",
+  },
+  mode: {
+    onlineTitle: "Jouer en ligne",
+    onlineDesc: "Connectez-vous au royaume. Vos personnages vivent sur le serveur et vous partagez le monde avec tous les autres joueurs connectés.",
+    onlineAria: "Jouer en ligne: connectez-vous au royaume persistant partagé",
+    offlineTitle: "Jouer hors ligne",
+    offlineDesc: "Un monde solo instantané dans votre navigateur. Rien n'est sauvegardé: parfait pour un combat rapide ou des tests.",
+    offlineAria: "Jouer hors ligne: démarrer une session solo locale instantanée",
+    tipTitle: "CONSEIL:",
+    tipText: "Pour une expérience optimale, désactivez les extensions de blocage de publicité sur ce site. Certains bloqueurs peuvent causer des latences.",
+  },
+  auth: {
+    enterRealm: "Entrer dans le royaume",
+    username: "Nom d'utilisateur",
+    usernameError: "Veuillez saisir votre nom d'utilisateur.",
+    usernamePlaceholder: "Saisir le nom d'utilisateur",
+    password: "Mot de passe",
+    passwordError: "Veuillez saisir votre mot de passe.",
+    passwordPlaceholder: "Saisir le mot de passe",
+    showPassword: "Afficher le mot de passe",
+    hidePassword: "Masquer le mot de passe",
+    logIn: "Se connecter",
+    createAccount: "Créer un compte",
+    back: "Retour",
+    realmList: "Liste des royaumes",
+    loadingRealms: "Chargement des royaumes...",
+    changeRealm: "Changer de royaume",
+    characters: "Personnages:",
+    createCharacter: "Créer un personnage",
+    characterName: "Nom du personnage",
+    characterNamePlaceholder: "Nom du personnage",
+    enterWorld: "Entrer dans le monde",
+    offlineCharacter: "Personnage hors ligne",
+    create: "Créer",
+  },
+  classes: {
+    warrior: "Guerrier",
+    paladin: "Paladin",
+    hunter: "Chasseur",
+    rogue: "Voleur",
+    priest: "Prêtre",
+    shaman: "Chaman",
+    mage: "Mage",
+    warlock: "Démoniste",
+    druid: "Druide",
+    warriorAria: "Classe Guerrier",
+    paladinAria: "Classe Paladin",
+    hunterAria: "Classe Chasseur",
+    rogueAria: "Classe Voleur",
+    priestAria: "Classe Prêtre",
+    shamanAria: "Classe Chaman",
+    mageAria: "Classe Mage",
+    warlockAria: "Classe Démoniste",
+    druidAria: "Classe Druide",
+  },
+  controls: {
+    title: "Guide des contrôles",
+    movement: "Déplacement",
+    moveTurn: "Avancer/Tourner",
+    strafe: "Déplacement latéral Gauche/Droite",
+    jump: "Sauter",
+    autorun: "Activer la course automatique",
+    combat: "Combat et interaction",
+    target: "Cibler l'ennemi",
+    spells: "Lancer des sorts",
+    interact: "Interagir/Fouiller",
+    nameplates: "Afficher les plaques de nom",
+    camera: "Caméra et souris",
+    rightDrag: "Glisser-clic droit",
+    leftDrag: "Glisser-clic gauche",
+    mouseWheel: "Molette de la souris",
+    mouselook: "Vue à la souris",
+    orbit: "Caméra orbitale",
+    zoom: "Zoom",
+    interfaces: "Interfaces",
+    charPane: "Fiche de personnage",
+    spellbook: "Grimoire",
+    questLog: "Journal de quêtes",
+    worldMap: "Carte du monde",
+    bags: "Inventaire des sacs",
+    emoteWheel: "Maintenir la roue des emotes",
+    friends: "Amis et guilde",
+    chat: "Ouvrir le chat",
+  },
+  ...shellStrings.fr_CA,
+  ...hudStrings.fr_CA,
+  ...abilityStrings.fr_CA,
+  ...questStrings.fr_CA,
+  ...itemStrings.fr_CA,
+  ...classAbilityNames.fr_CA,
+  ...itemNames.fr_CA,
+  ...worldNames.fr_CA,
+  ...mergeStrings.fr_CA,
+  entities: {
+    ...itemNames.fr_CA.entities,
+    ...worldNames.fr_CA.entities,
+    abilities: { ...itemNames.fr_CA.entities.abilities, ...mergeExtra.fr_CA.abilities },
+    items: { ...itemNames.fr_CA.entities.items, ...mergeEntities.fr_CA.items, ...mergeExtra.fr_CA.items },
+    mobs: { ...worldNames.fr_CA.entities.mobs, ...mergeEntities.fr_CA.mobs, ...mergeExtra.fr_CA.mobs },
+    npcs: { ...worldNames.fr_CA.entities.npcs, ...mergeExtra.fr_CA.npcs },
+    quests: { ...worldNames.fr_CA.entities.quests, ...mergeEntities.fr_CA.quests, ...mergeExtra.fr_CA.quests },
+    dungeons: { ...worldNames.fr_CA.entities.dungeons, ...mergeExtra.fr_CA.dungeons },
+  },
+};
