+++ src/ui/i18n.locales/it_IT.ts
@@ -1,0 +1,169 @@
+import {
+  en,
+  shellStrings,
+  hudStrings,
+  abilityStrings,
+  questStrings,
+  itemStrings,
+  classAbilityNames,
+  itemNames,
+  mergeStrings,
+  mergeEntities,
+  mergeExtra,
+  gameStringsItIT,
+} from '../i18n.en';
+import { worldEntityText as worldNames } from '../world_entity_i18n';
+
+export const it_IT: typeof en = {
+  meta: { builtOn: "Compilato il {date}" },
+  realmTypes: { normal: "Normale", pvp: "PvP", rp: "GdR", rpPvp: "GdR-PvP" },
+  game: gameStringsItIT,
+  nav: {
+    home: "Pagina principale",
+    play: "Gioca",
+    stats: "Statistiche",
+    about: "Informazioni",
+    highscores: "Classifiche",
+    wiki: "Wiki",
+    news: "Novità",
+    download: "Scarica",
+    loginRegister: "Accedi/Registrati",
+    donate: "Dona",
+  },
+  stats: {
+    title: "Stato del Reame",
+    accountsCreated: "Account creati",
+    playersOnline: "Giocatori online",
+    realmName: "Nome del reame",
+  },
+  footer: {
+    copyright: "2026 World of ClaudeCraft",
+    githubLink: "https://github.com/levy-street/world-of-claudecraft",
+    githubLabel: "Progetto open source",
+    terms: "Termini di servizio",
+    privacy: "Informativa sulla privacy",
+    discordLabel: "Entra nel Discord",
+  },
+  highscores: {
+    title: "Classifica dei punteggi più alti",
+    desc: "Segui i più grandi campioni del reame e confronta i tuoi progressi.",
+  },
+  wiki: {
+    title: "Wiki e guida del gioco",
+    desc: "Scopri i segreti del reame, le guide di classe e le strategie.",
+  },
+  news: {
+    title: "Notizie e aggiornamenti",
+    desc: "Leggi le ultime note sulla patch, gli eventi e gli aggiornamenti della community.",
+  },
+  download: {
+    title: "Scarica il launcher per desktop",
+    desc: "Ottieni il launcher autonomo per prestazioni ottimizzate e gioco a schermo intero.",
+  },
+  comingSoon: {
+    placeholder: "Prossimamente...",
+    featureComingSoon: "Questa funzione sarà presto disponibile nel reame.",
+  },
+  mode: {
+    onlineTitle: "Gioca online",
+    onlineDesc: "Accedi al reame. I tuoi personaggi vivono sul server e condividi il mondo con tutti gli altri giocatori connessi.",
+    onlineAria: "Gioca online: accedi al reame condiviso persistente",
+    offlineTitle: "Gioca offline",
+    offlineDesc: "Un mondo a giocatore singolo istantaneo nel tuo browser. Nulla viene salvato: perfetto per un combattimento rapido o per testare.",
+    offlineAria: "Gioca offline: avvia una sessione locale a giocatore singolo istantanea",
+    tipTitle: "SUGGERIMENTO:",
+    tipText: "Per un'esperienza ottimale, disattiva le estensioni di blocco della pubblicità su questo sito. Alcuni utenti segnalano che i blocchi possono causare rallentamenti.",
+  },
+  auth: {
+    enterRealm: "Entra nel reame",
+    username: "Nome utente",
+    usernameError: "Inserisci il tuo nome utente.",
+    usernamePlaceholder: "Inserisci il nome utente",
+    password: "Password",
+    passwordError: "Inserisci la tua password.",
+    passwordPlaceholder: "Inserisci la password",
+    showPassword: "Mostra password",
+    hidePassword: "Nascondi password",
+    logIn: "Accedi",
+    createAccount: "Crea account",
+    back: "Indietro",
+    realmList: "Lista dei reami",
+    loadingRealms: "Caricamento dei reami...",
+    changeRealm: "Cambia reame",
+    characters: "Personaggi:",
+    createCharacter: "Crea personaggio",
+    characterName: "Nome del personaggio",
+    characterNamePlaceholder: "Nome del personaggio",
+    enterWorld: "Entra nel mondo",
+    offlineCharacter: "Personaggio offline",
+    create: "Crea",
+  },
+  classes: {
+    warrior: "Guerriero",
+    paladin: "Paladino",
+    hunter: "Cacciatore",
+    rogue: "Ladro",
+    priest: "Sacerdote",
+    shaman: "Sciamano",
+    mage: "Mago",
+    warlock: "Stregone",
+    druid: "Druido",
+    warriorAria: "Classe Guerriero",
+    paladinAria: "Classe Paladino",
+    hunterAria: "Classe Cacciatore",
+    rogueAria: "Classe Ladro",
+    priestAria: "Classe Sacerdote",
+    shamanAria: "Classe Sciamano",
+    mageAria: "Classe Mago",
+    warlockAria: "Classe Stregone",
+    druidAria: "Classe Druido",
+  },
+  controls: {
+    title: "Guida ai controlli",
+    movement: "Movimento",
+    moveTurn: "Muoviti/Gira",
+    strafe: "Passi laterali Sinistra/Destra",
+    jump: "Salta",
+    autorun: "Attiva corsa automatica",
+    combat: "Combattimento e interazione",
+    target: "Seleziona nemico come bersaglio",
+    spells: "Lancia incantesimi",
+    interact: "Interagisci/Saccheggia",
+    nameplates: "Attiva targhette dei nomi",
+    camera: "Telecamera e mouse",
+    rightDrag: "Trascina con il tasto destro",
+    leftDrag: "Trascina con il tasto sinistro",
+    mouseWheel: "Rotella del mouse",
+    mouselook: "Visuale con il mouse",
+    orbit: "Orbita telecamera",
+    zoom: "Zoom",
+    interfaces: "Interfacce",
+    charPane: "Schermata personaggio",
+    spellbook: "Libro degli incantesimi",
+    questLog: "Registro delle missioni",
+    worldMap: "Mappa del mondo",
+    bags: "Inventario borse",
+    emoteWheel: "Tieni premuta la ruota emote",
+    friends: "Amici e gilda",
+    chat: "Apri chat",
+  },
+  ...shellStrings.it_IT,
+  ...hudStrings.it_IT,
+  ...abilityStrings.it_IT,
+  ...questStrings.it_IT,
+  ...itemStrings.it_IT,
+  ...classAbilityNames.it_IT,
+  ...itemNames.it_IT,
+  ...worldNames.it_IT,
+  ...mergeStrings.it_IT,
+  entities: {
+    ...itemNames.it_IT.entities,
+    ...worldNames.it_IT.entities,
+    abilities: { ...itemNames.it_IT.entities.abilities, ...mergeExtra.it_IT.abilities },
+    items: { ...itemNames.it_IT.entities.items, ...mergeEntities.it_IT.items, ...mergeExtra.it_IT.items },
+    mobs: { ...worldNames.it_IT.entities.mobs, ...mergeEntities.it_IT.mobs, ...mergeExtra.it_IT.mobs },
+    npcs: { ...worldNames.it_IT.entities.npcs, ...mergeExtra.it_IT.npcs },
+    quests: { ...worldNames.it_IT.entities.quests, ...mergeEntities.it_IT.quests, ...mergeExtra.it_IT.quests },
+    dungeons: { ...worldNames.it_IT.entities.dungeons, ...mergeExtra.it_IT.dungeons },
+  },
+};
