+++ src/ui/i18n.locales/ja_JP.ts
@@ -1,0 +1,169 @@
+import {
+  en,
+  shellStrings,
+  hudStrings,
+  abilityStrings,
+  questStrings,
+  itemStrings,
+  classAbilityNames,
+  itemNames,
+  mergeStrings,
+  mergeEntities,
+  mergeExtra,
+  gameStringsJaJP,
+} from '../i18n.en';
+import { worldEntityText as worldNames } from '../world_entity_i18n';
+
+export const ja_JP: typeof en = {
+  meta: { builtOn: "ビルド日：{date}" },
+  realmTypes: { normal: "ノーマル", pvp: "PvP", rp: "RP", rpPvp: "RP-PvP" },
+  game: gameStringsJaJP,
+  nav: {
+    home: "ホーム",
+    play: "プレイ",
+    stats: "統計",
+    about: "情報",
+    highscores: "ハイスコア",
+    wiki: "Wiki",
+    news: "ニュース",
+    download: "ダウンロード",
+    loginRegister: "ログイン/登録",
+    donate: "寄付",
+  },
+  stats: {
+    title: "レルムステータス",
+    accountsCreated: "作成アカウント数",
+    playersOnline: "オンラインプレイヤー数",
+    realmName: "レルム名",
+  },
+  footer: {
+    copyright: "2026 World of ClaudeCraft",
+    githubLink: "https://github.com/levy-street/world-of-claudecraft",
+    githubLabel: "オープンソースプロジェクト",
+    terms: "利用規約",
+    privacy: "プライバシーポリシー",
+    discordLabel: "Discordに参加する",
+  },
+  highscores: {
+    title: "ハイスコアリーダーボード",
+    desc: "レルムの偉大なチャンピオンたちを確認し、あなたの進行度と比較しましょう。",
+  },
+  wiki: {
+    title: "ゲームWiki & ガイド",
+    desc: "レルムの秘密、クラスガイド、戦略を確認しましょう。",
+  },
+  news: {
+    title: "ニュース & アップデート",
+    desc: "最新のパッチノート、イベント、コミュニティの更新情報を確認しましょう。",
+  },
+  download: {
+    title: "デスクトップランチャーのダウンロード",
+    desc: "最適化されたパフォーマンスと全画面表示でのプレイのために、スタンドアロン版ランチャーを入手してください。",
+  },
+  comingSoon: {
+    placeholder: "近日公開...",
+    featureComingSoon: "この機能はまもなくレルムに実装されます。",
+  },
+  mode: {
+    onlineTitle: "オンラインでプレイ",
+    onlineDesc: "レルムにログインします。キャラクターはサーバー上に保存され、オンラインの他の全プレイヤーと同じ世界を共有します。",
+    onlineAria: "オンラインでプレイ：永続的な共有レルムにログインします",
+    offlineTitle: "オフラインでプレイ",
+    offlineDesc: "ブラウザ上で即座にシングルプレイヤーの世界を起動します。保存は一切されず、小規模な対戦やテストに最適です。",
+    offlineAria: "オフラインでプレイ：即座にローカルシングルプレイヤーセッションを開始します",
+    tipTitle: "ヒント：",
+    tipText: "よりスムーズな体験のために、本サイトでの広告ブロック拡張機能を無効にしてください。コミュニティの報告によると、一部のブロック機能がラグの原因になる場合があります。",
+  },
+  auth: {
+    enterRealm: "レルムに入る",
+    username: "ユーザー名",
+    usernameError: "ユーザー名を入力してください。",
+    usernamePlaceholder: "ユーザー名を入力",
+    password: "パスワード",
+    passwordError: "パスワードを入力してください。",
+    passwordPlaceholder: "パスワードを入力",
+    showPassword: "パスワードを表示",
+    hidePassword: "パスワードを非表示",
+    logIn: "ログイン",
+    createAccount: "アカウント作成",
+    back: "戻る",
+    realmList: "レルムリスト",
+    loadingRealms: "レルムを読み込み中...",
+    changeRealm: "レルム変更",
+    characters: "キャラクター数：",
+    createCharacter: "キャラクター作成",
+    characterName: "キャラクター名",
+    characterNamePlaceholder: "キャラクター名",
+    enterWorld: "世界に入る",
+    offlineCharacter: "オフラインキャラクター",
+    create: "作成",
+  },
+  classes: {
+    warrior: "戦士",
+    paladin: "パラディン",
+    hunter: "ハンター",
+    rogue: "ローグ",
+    priest: "プリースト",
+    shaman: "シャーマン",
+    mage: "メイジ",
+    warlock: "ウォーロック",
+    druid: "ドルイド",
+    warriorAria: "戦士クラス",
+    paladinAria: "パラディンクラス",
+    hunterAria: "ハンタークラス",
+    rogueAria: "ローグクラス",
+    priestAria: "プリーストクラス",
+    shamanAria: "シャーマンクラス",
+    mageAria: "メイジクラス",
+    warlockAria: "ウォーロッククラス",
+    druidAria: "ドルイドクラス",
+  },
+  controls: {
+    title: "操作ガイド",
+    movement: "移動",
+    moveTurn: "移動 / 旋回",
+    strafe: "左右平行移動",
+    jump: "ジャンプ",
+    autorun: "オートラン切り替え",
+    combat: "戦闘 & 相互作用",
+    target: "敵をターゲット",
+    spells: "呪文の詠唱",
+    interact: "相互作用 / 戦利品回収",
+    nameplates: "ネームプレート表示切り替え",
+    camera: "カメラ & マウス",
+    rightDrag: "右ドラッグ",
+    leftDrag: "左ドラッグ",
+    mouseWheel: "マウスホイール",
+    mouselook: "マウスルック",
+    orbit: "カメラの旋回",
+    zoom: "ズーム",
+    interfaces: "画面表示",
+    charPane: "キャラクター画面",
+    spellbook: "魔道書",
+    questLog: "クエストログ",
+    worldMap: "ワールドマップ",
+    bags: "バッグのインベントリ",
+    emoteWheel: "エモートホイールを長押し",
+    friends: "フレンド & ギルド",
+    chat: "チャットを開く",
+  },
+  ...shellStrings.ja_JP,
+  ...hudStrings.ja_JP,
+  ...abilityStrings.ja_JP,
+  ...questStrings.ja_JP,
+  ...itemStrings.ja_JP,
+  ...classAbilityNames.ja_JP,
+  ...itemNames.ja_JP,
+  ...worldNames.ja_JP,
+  ...mergeStrings.ja_JP,
+  entities: {
+    ...itemNames.ja_JP.entities,
+    ...worldNames.ja_JP.entities,
+    abilities: { ...itemNames.ja_JP.entities.abilities, ...mergeExtra.ja_JP.abilities },
+    items: { ...itemNames.ja_JP.entities.items, ...mergeEntities.ja_JP.items, ...mergeExtra.ja_JP.items },
+    mobs: { ...worldNames.ja_JP.entities.mobs, ...mergeEntities.ja_JP.mobs, ...mergeExtra.ja_JP.mobs },
+    npcs: { ...worldNames.ja_JP.entities.npcs, ...mergeExtra.ja_JP.npcs },
+    quests: { ...worldNames.ja_JP.entities.quests, ...mergeEntities.ja_JP.quests, ...mergeExtra.ja_JP.quests },
+    dungeons: { ...worldNames.ja_JP.entities.dungeons, ...mergeExtra.ja_JP.dungeons },
+  },
+};
