+++ src/ui/i18n.locales/ko_KR.ts
@@ -1,0 +1,169 @@
+import {
+  en,
+  shellStrings,
+  hudStrings,
+  abilityStrings,
+  questStrings,
+  itemStrings,
+  classAbilityNames,
+  itemNames,
+  mergeStrings,
+  mergeEntities,
+  mergeExtra,
+  gameStringsKoKR,
+} from '../i18n.en';
+import { worldEntityText as worldNames } from '../world_entity_i18n';
+
+export const ko_KR: typeof en = {
+  meta: { builtOn: "{date} 빌드" },
+  realmTypes: { normal: "일반", pvp: "PvP", rp: "롤플레잉", rpPvp: "롤플레잉 PvP" },
+  game: gameStringsKoKR,
+  nav: {
+    home: "홈",
+    play: "플레이",
+    stats: "통계",
+    about: "정보",
+    highscores: "순위표",
+    wiki: "위키",
+    news: "새소식",
+    download: "다운로드",
+    loginRegister: "로그인/회원가입",
+    donate: "후원",
+  },
+  stats: {
+    title: "렐름 상태",
+    accountsCreated: "생성된 계정 수",
+    playersOnline: "현재 접속자 수",
+    realmName: "렐름 이름",
+  },
+  footer: {
+    copyright: "2026 World of ClaudeCraft",
+    githubLink: "https://github.com/levy-street/world-of-claudecraft",
+    githubLabel: "오픈 소스 프로젝트",
+    terms: "서비스 이용약관",
+    privacy: "개인정보 처리방침",
+    discordLabel: "Discord 참여하기",
+  },
+  highscores: {
+    title: "최고 점수 순위표",
+    desc: "렐름의 가장 위대한 챔피언들을 확인하고 당신의 진행 상태를 비교해 보세요.",
+  },
+  wiki: {
+    title: "게임 위키 및 가이드",
+    desc: "렐름의 비밀과 클래스 가이드, 전략 등을 확인해 보세요.",
+  },
+  news: {
+    title: "새소식 및 업데이트",
+    desc: "최신 패치 노트, 이벤트 소식, 커뮤니티 업데이트를 확인해 보세요.",
+  },
+  download: {
+    title: "데스크톱 런처 다운로드",
+    desc: "최적화된 성능과 전체 화면 플레이를 위해 독립형 런처를 다운로드하세요.",
+  },
+  comingSoon: {
+    placeholder: "준비 중...",
+    featureComingSoon: "이 기능은 곧 렐름에 업데이트될 예정입니다.",
+  },
+  mode: {
+    onlineTitle: "온라인 플레이",
+    onlineDesc: "렐름에 로그인합니다. 생성한 캐릭터는 서버에 저장되며 온라인에 있는 다른 모든 플레이어와 같은 세상을 공유합니다.",
+    onlineAria: "온라인 플레이: 지속적으로 공유되는 렐름에 로그인합니다",
+    offlineTitle: "오프라인 플레이",
+    offlineDesc: "브라우저에서 즉시 일인용 세상을 플레이합니다. 아무것도 저장되지 않아 빠른 전투나 연습에 적합합니다.",
+    offlineAria: "오프라인 플레이: 즉각적인 로컬 싱글 플레이 세션을 시작합니다",
+    tipTitle: "팁:",
+    tipText: "가장 원활한 이용을 위해 이 웹사이트의 광고 차단기 확장 프로그램을 비활성화해 주세요. 일부 차단기가 게임 성능 지연을 유발할 수 있습니다.",
+  },
+  auth: {
+    enterRealm: "렐름 입장",
+    username: "사용자 이름",
+    usernameError: "사용자 이름을 입력해 주세요.",
+    usernamePlaceholder: "사용자 이름 입력",
+    password: "비밀번호",
+    passwordError: "비밀번호를 입력해 주세요.",
+    passwordPlaceholder: "비밀번호 입력",
+    showPassword: "비밀번호 표시",
+    hidePassword: "비밀번호 숨기기",
+    logIn: "로그인",
+    createAccount: "계정 생성",
+    back: "뒤로 가기",
+    realmList: "렐름 목록",
+    loadingRealms: "렐름을 불러오는 중...",
+    changeRealm: "렐름 변경",
+    characters: "캐릭터 목록:",
+    createCharacter: "캐릭터 생성",
+    characterName: "캐릭터 이름",
+    characterNamePlaceholder: "캐릭터 이름",
+    enterWorld: "세계 입장",
+    offlineCharacter: "오프라인 캐릭터",
+    create: "생성",
+  },
+  classes: {
+    warrior: "전사",
+    paladin: "성기사",
+    hunter: "사냥꾼",
+    rogue: "도적",
+    priest: "사제",
+    shaman: "주술사",
+    mage: "마법사",
+    warlock: "흑마법사",
+    druid: "드루이드",
+    warriorAria: "전사 직업",
+    paladinAria: "성기사 직업",
+    hunterAria: "사냥꾼 직업",
+    rogueAria: "도적 직업",
+    priestAria: "사제 직업",
+    shamanAria: "주술사 직업",
+    mageAria: "마법사 직업",
+    warlockAria: "흑마법사 직업",
+    druidAria: "드루이드 직업",
+  },
+  controls: {
+    title: "조작 가이드",
+    movement: "이동 조작",
+    moveTurn: "이동 / 회전",
+    strafe: "좌/우 평행 이동",
+    jump: "점프",
+    autorun: "자동 달리기 전환",
+    combat: "전투 및 상호작용",
+    target: "적 대상 지정",
+    spells: "주문 시전",
+    interact: "상호작용 / 전리품 획득",
+    nameplates: "이름표 전환",
+    camera: "카메라 및 마우스",
+    rightDrag: "우클릭 드래그",
+    leftDrag: "좌클릭 드래그",
+    mouseWheel: "마우스 휠",
+    mouselook: "마우스 시점",
+    orbit: "카메라 회전",
+    zoom: "줌",
+    interfaces: "인터페이스",
+    charPane: "캐릭터 정보",
+    spellbook: "마법책",
+    questLog: "퀘스트 목록",
+    worldMap: "세계 지도",
+    bags: "가방 소지품",
+    emoteWheel: "감정표현 휠 길게 누르기",
+    friends: "친구 및 길드",
+    chat: "채팅창 열기",
+  },
+  ...shellStrings.ko_KR,
+  ...hudStrings.ko_KR,
+  ...abilityStrings.ko_KR,
+  ...questStrings.ko_KR,
+  ...itemStrings.ko_KR,
+  ...classAbilityNames.ko_KR,
+  ...itemNames.ko_KR,
+  ...worldNames.ko_KR,
+  ...mergeStrings.ko_KR,
+  entities: {
+    ...itemNames.ko_KR.entities,
+    ...worldNames.ko_KR.entities,
+    abilities: { ...itemNames.ko_KR.entities.abilities, ...mergeExtra.ko_KR.abilities },
+    items: { ...itemNames.ko_KR.entities.items, ...mergeEntities.ko_KR.items, ...mergeExtra.ko_KR.items },
+    mobs: { ...worldNames.ko_KR.entities.mobs, ...mergeEntities.ko_KR.mobs, ...mergeExtra.ko_KR.mobs },
+    npcs: { ...worldNames.ko_KR.entities.npcs, ...mergeExtra.ko_KR.npcs },
+    quests: { ...worldNames.ko_KR.entities.quests, ...mergeEntities.ko_KR.quests, ...mergeExtra.ko_KR.quests },
+    dungeons: { ...worldNames.ko_KR.entities.dungeons, ...mergeExtra.ko_KR.dungeons },
+  },
+};
