+++ src/ui/i18n.locales/pt_BR.ts
@@ -1,0 +1,169 @@
+import {
+  en,
+  shellStrings,
+  hudStrings,
+  abilityStrings,
+  questStrings,
+  itemStrings,
+  classAbilityNames,
+  itemNames,
+  mergeStrings,
+  mergeEntities,
+  mergeExtra,
+  gameStringsPtBR,
+} from '../i18n.en';
+import { worldEntityText as worldNames } from '../world_entity_i18n';
+
+export const pt_BR: typeof en = {
+  meta: { builtOn: "Compilado em {date}" },
+  realmTypes: { normal: "Normal", pvp: "PvP", rp: "RP", rpPvp: "RP-PvP" },
+  game: gameStringsPtBR,
+  nav: {
+    home: "Início",
+    play: "Jogar",
+    stats: "Estatísticas",
+    about: "Sobre",
+    highscores: "Recordes",
+    wiki: "Wiki",
+    news: "Notícias",
+    download: "Baixar",
+    loginRegister: "Entrar/Registrar",
+    donate: "Doar",
+  },
+  stats: {
+    title: "Status do Reino",
+    accountsCreated: "Contas criadas",
+    playersOnline: "Jogadores online",
+    realmName: "Nome do reino",
+  },
+  footer: {
+    copyright: "2026 World of ClaudeCraft",
+    githubLink: "https://github.com/levy-street/world-of-claudecraft",
+    githubLabel: "Projeto de código aberto",
+    terms: "Termos de serviço",
+    privacy: "Política de privacidade",
+    discordLabel: "Juntar-se ao Discord",
+  },
+  highscores: {
+    title: "Tabela de recordes",
+    desc: "Acompanhe os maiores campeões do reino e compare o seu progresso.",
+  },
+  wiki: {
+    title: "Wiki e guia do jogo",
+    desc: "Descubra os segredos do reino, guias de classes e estratégias.",
+  },
+  news: {
+    title: "Notícias e atualizações",
+    desc: "Leia as últimas notas do patch, eventos e atualizações da comunidade.",
+  },
+  download: {
+    title: "Baixar launcher para desktop",
+    desc: "Obtenha o launcher independente para desempenho otimizado e jogo em tela cheia.",
+  },
+  comingSoon: {
+    placeholder: "Em breve...",
+    featureComingSoon: "Este recurso estará disponível no reino em breve.",
+  },
+  mode: {
+    onlineTitle: "Jogar online",
+    onlineDesc: "Entre no reino. Seus personagens vivem no servidor e você compartilha o mundo com todos os outros online.",
+    onlineAria: "Jogar online: entre no reino compartilhado persistente",
+    offlineTitle: "Jogar offline",
+    offlineDesc: "Mundo solo instantâneo no seu navegador. Nada é salvo: perfeito para uma luta rápida ou testes.",
+    offlineAria: "Jogar offline: inicie uma sessão local de um jogador instantânea",
+    tipTitle: "DICA:",
+    tipText: "Para uma experiência mais suave, desative os bloqueadores de anúncios neste site. Relatos da comunidade indicam que alguns bloqueadores podem causar lag.",
+  },
+  auth: {
+    enterRealm: "Entrar no reino",
+    username: "Nome de usuário",
+    usernameError: "Por favor, insira seu nome de usuário.",
+    usernamePlaceholder: "Insira o usuário",
+    password: "Senha",
+    passwordError: "Por favor, insira sua senha.",
+    passwordPlaceholder: "Insira a senha",
+    showPassword: "Mostrar senha",
+    hidePassword: "Ocultar senha",
+    logIn: "Entrar",
+    createAccount: "Criar conta",
+    back: "Voltar",
+    realmList: "Lista de reinos",
+    loadingRealms: "Carregando reinos...",
+    changeRealm: "Mudar de reino",
+    characters: "Personagens:",
+    createCharacter: "Criar personagem",
+    characterName: "Nome do personagem",
+    characterNamePlaceholder: "Nome do personagem",
+    enterWorld: "Entrar no mundo",
+    offlineCharacter: "Personagem offline",
+    create: "Criar",
+  },
+  classes: {
+    warrior: "Guerreiro",
+    paladin: "Paladino",
+    hunter: "Caçador",
+    rogue: "Ladino",
+    priest: "Sacerdote",
+    shaman: "Xamã",
+    mage: "Mago",
+    warlock: "Bruxo",
+    druid: "Druida",
+    warriorAria: "Classe Guerreiro",
+    paladinAria: "Classe Paladino",
+    hunterAria: "Classe Caçador",
+    rogueAria: "Classe Ladino",
+    priestAria: "Classe Sacerdote",
+    shamanAria: "Classe Xamã",
+    mageAria: "Classe Mago",
+    warlockAria: "Classe Bruxo",
+    druidAria: "Classe Druida",
+  },
+  controls: {
+    title: "Guia de controles",
+    movement: "Movimentação",
+    moveTurn: "Mover / Virar",
+    strafe: "Passo lateral Esquerda/Direita",
+    jump: "Pular",
+    autorun: "Alternar corrida automática",
+    combat: "Combate e interação",
+    target: "Selecionar inimigo como alvo",
+    spells: "Conjurar feitiços",
+    interact: "Interagir / Saquear",
+    nameplates: "Alternar placas de nome",
+    camera: "Câmera e mouse",
+    rightDrag: "Arrastar com botão direito",
+    leftDrag: "Arrastar com botão esquerdo",
+    mouseWheel: "Roda do mouse",
+    mouselook: "Olhar com o mouse",
+    orbit: "Orbitar câmera",
+    zoom: "Zoom",
+    interfaces: "Interfaces",
+    charPane: "Painel do personagem",
+    spellbook: "Livro de feitiços",
+    questLog: "Registro de missões",
+    worldMap: "Mapa do mundo",
+    bags: "Inventário de bolsas",
+    emoteWheel: "Segurar roda de emotes",
+    friends: "Amigos e guilda",
+    chat: "Abrir chat",
+  },
+  ...shellStrings.pt_BR,
+  ...hudStrings.pt_BR,
+  ...abilityStrings.pt_BR,
+  ...questStrings.pt_BR,
+  ...itemStrings.pt_BR,
+  ...classAbilityNames.pt_BR,
+  ...itemNames.pt_BR,
+  ...worldNames.pt_BR,
+  ...mergeStrings.pt_BR,
+  entities: {
+    ...itemNames.pt_BR.entities,
+    ...worldNames.pt_BR.entities,
+    abilities: { ...itemNames.pt_BR.entities.abilities, ...mergeExtra.pt_BR.abilities },
+    items: { ...itemNames.pt_BR.entities.items, ...mergeEntities.pt_BR.items, ...mergeExtra.pt_BR.items },
+    mobs: { ...worldNames.pt_BR.entities.mobs, ...mergeEntities.pt_BR.mobs, ...mergeExtra.pt_BR.mobs },
+    npcs: { ...worldNames.pt_BR.entities.npcs, ...mergeExtra.pt_BR.npcs },
+    quests: { ...worldNames.pt_BR.entities.quests, ...mergeEntities.pt_BR.quests, ...mergeExtra.pt_BR.quests },
+    dungeons: { ...worldNames.pt_BR.entities.dungeons, ...mergeExtra.pt_BR.dungeons },
+  },
+};
