+++ src/ui/i18n.locales/ru_RU.ts
@@ -1,0 +1,169 @@
+import {
+  en,
+  shellStrings,
+  hudStrings,
+  abilityStrings,
+  questStrings,
+  itemStrings,
+  classAbilityNames,
+  itemNames,
+  mergeStrings,
+  mergeEntities,
+  mergeExtra,
+  gameStringsRuRU,
+} from '../i18n.en';
+import { worldEntityText as worldNames } from '../world_entity_i18n';
+
+export const ru_RU: typeof en = {
+  meta: { builtOn: "Сборка от {date}" },
+  realmTypes: { normal: "Обычный", pvp: "PvP", rp: "Ролевой", rpPvp: "Ролевой-PvP" },
+  game: gameStringsRuRU,
+  nav: {
+    home: "Главная",
+    play: "Играть",
+    stats: "Статистика",
+    about: "О проекте",
+    highscores: "Таблицы рекордов",
+    wiki: "Вики",
+    news: "Новости",
+    download: "Скачать",
+    loginRegister: "Войти / Зарегистрироваться",
+    donate: "Пожертвовать",
+  },
+  stats: {
+    title: "Статус игрового мира",
+    accountsCreated: "Создано учетных записей",
+    playersOnline: "Игроков в сети",
+    realmName: "Название игрового мира",
+  },
+  footer: {
+    copyright: "2026 World of ClaudeCraft",
+    githubLink: "https://github.com/levy-street/world-of-claudecraft",
+    githubLabel: "Проект с открытым исходным кодом",
+    terms: "Условия использования",
+    privacy: "Политика конфиденциальности",
+    discordLabel: "Присоединиться к Discord",
+  },
+  highscores: {
+    title: "Таблица рекордов",
+    desc: "Отслеживайте величайших чемпионов игрового мира и сравнивайте свои успехи.",
+  },
+  wiki: {
+    title: "Вики и руководство",
+    desc: "Откройте секреты игрового мира, руководства по классам и стратегии.",
+  },
+  news: {
+    title: "Новости и обновления",
+    desc: "Читайте описание последних обновлений, информацию о событиях и новости сообщества.",
+  },
+  download: {
+    title: "Скачать загрузчик для ПК",
+    desc: "Загрузите отдельный клиент для оптимальной производительности и игры на весь экран.",
+  },
+  comingSoon: {
+    placeholder: "Скоро...",
+    featureComingSoon: "Эта функция скоро появится в игре.",
+  },
+  mode: {
+    onlineTitle: "Играть по сети",
+    onlineDesc: "Войдите в игровой мир. Ваши персонажи сохраняются на сервере, а вы делите игровой мир с другими игроками в сети.",
+    onlineAria: "Играть по сети: вход в постоянный общий игровой мир",
+    offlineTitle: "Играть вне сети",
+    offlineDesc: "Мгновенный одиночный мир прямо в браузере. Прогресс не сохраняется: идеально для быстрой драки или тестирования.",
+    offlineAria: "Играть вне сети: запуск мгновенного локального одиночного сеанса",
+    tipTitle: "СОВЕТ:",
+    tipText: "Для максимально плавной игры отключите блокировщики рекламы на этом сайте. По отзывам сообщества, некоторые блокировщики могут вызывать задержки.",
+  },
+  auth: {
+    enterRealm: "Войти в игровой мир",
+    username: "Имя пользователя",
+    usernameError: "Пожалуйста, введите имя пользователя.",
+    usernamePlaceholder: "Введите имя пользователя",
+    password: "Пароль",
+    passwordError: "Пожалуйста, введите пароль.",
+    passwordPlaceholder: "Введите пароль",
+    showPassword: "Показать пароль",
+    hidePassword: "Скрыть пароль",
+    logIn: "Войти",
+    createAccount: "Создать учетную запись",
+    back: "Назад",
+    realmList: "Список игровых миров",
+    loadingRealms: "Загрузка списка миров...",
+    changeRealm: "Сменить игровой мир",
+    characters: "Персонажи:",
+    createCharacter: "Создать персонажа",
+    characterName: "Имя персонажа",
+    characterNamePlaceholder: "Имя персонажа",
+    enterWorld: "Войти в мир",
+    offlineCharacter: "Персонаж вне сети",
+    create: "Создать",
+  },
+  classes: {
+    warrior: "Воин",
+    paladin: "Паладин",
+    hunter: "Охотник",
+    rogue: "Разбойник",
+    priest: "Жрец",
+    shaman: "Шаман",
+    mage: "Маг",
+    warlock: "Чернокнижник",
+    druid: "Друид",
+    warriorAria: "Класс Воин",
+    paladinAria: "Класс Паладин",
+    hunterAria: "Класс Охотник",
+    rogueAria: "Класс Разбойник",
+    priestAria: "Класс Жрец",
+    shamanAria: "Класс Шаман",
+    mageAria: "Класс Маг",
+    warlockAria: "Класс Чернокнижник",
+    druidAria: "Класс Друид",
+  },
+  controls: {
+    title: "Руководство по управлению",
+    movement: "Передвижение",
+    moveTurn: "Движение / поворот",
+    strafe: "Стрейф влево/вправо",
+    jump: "Прыжок",
+    autorun: "Вкл/выкл автобег",
+    combat: "Бой и взаимодействие",
+    target: "Выбор врага",
+    spells: "Заклинания",
+    interact: "Взаимодействие / сбор добычи",
+    nameplates: "Вкл/выкл индикаторы здоровья",
+    camera: "Камера и мышь",
+    rightDrag: "Перетаскивание правой кнопкой",
+    leftDrag: "Перетаскивание левой кнопкой",
+    mouseWheel: "Колесико мыши",
+    mouselook: "Обзор мышью",
+    orbit: "Вращение камеры",
+    zoom: "Приближение",
+    interfaces: "Интерфейсы",
+    charPane: "Окно персонажа",
+    spellbook: "Книга заклинаний",
+    questLog: "Журнал заданий",
+    worldMap: "Карта мира",
+    bags: "Сумки",
+    emoteWheel: "Удерживать колесо эмоций",
+    friends: "Друзья и гильдия",
+    chat: "Открыть чат",
+  },
+  ...shellStrings.ru_RU,
+  ...hudStrings.ru_RU,
+  ...abilityStrings.ru_RU,
+  ...questStrings.ru_RU,
+  ...itemStrings.ru_RU,
+  ...classAbilityNames.ru_RU,
+  ...itemNames.ru_RU,
+  ...worldNames.ru_RU,
+  ...mergeStrings.ru_RU,
+  entities: {
+    ...itemNames.ru_RU.entities,
+    ...worldNames.ru_RU.entities,
+    abilities: { ...itemNames.ru_RU.entities.abilities, ...mergeExtra.ru_RU.abilities },
+    items: { ...itemNames.ru_RU.entities.items, ...mergeEntities.ru_RU.items, ...mergeExtra.ru_RU.items },
+    mobs: { ...worldNames.ru_RU.entities.mobs, ...mergeEntities.ru_RU.mobs, ...mergeExtra.ru_RU.mobs },
+    npcs: { ...worldNames.ru_RU.entities.npcs, ...mergeExtra.ru_RU.npcs },
+    quests: { ...worldNames.ru_RU.entities.quests, ...mergeEntities.ru_RU.quests, ...mergeExtra.ru_RU.quests },
+    dungeons: { ...worldNames.ru_RU.entities.dungeons, ...mergeExtra.ru_RU.dungeons },
+  },
+};
