+++ src/ui/i18n.locales/zh_CN.ts
@@ -1,0 +1,169 @@
+import {
+  en,
+  shellStrings,
+  hudStrings,
+  abilityStrings,
+  questStrings,
+  itemStrings,
+  classAbilityNames,
+  itemNames,
+  mergeStrings,
+  mergeEntities,
+  mergeExtra,
+  gameStringsZhCN,
+} from '../i18n.en';
+import { worldEntityText as worldNames } from '../world_entity_i18n';
+
+export const zh_CN: typeof en = {
+  meta: { builtOn: "构建于 {date}" },
+  realmTypes: { normal: "普通", pvp: "PvP", rp: "RP", rpPvp: "RP-PvP" },
+  game: gameStringsZhCN,
+  nav: {
+    home: "首页",
+    play: "开始游戏",
+    stats: "数据统计",
+    about: "关于",
+    highscores: "排行榜",
+    wiki: "百科指南",
+    news: "新闻资讯",
+    download: "下载客户端",
+    loginRegister: "登录/注册",
+    donate: "捐赠",
+  },
+  stats: {
+    title: "服务器状态",
+    accountsCreated: "已创建账号",
+    playersOnline: "在线玩家数",
+    realmName: "服务器名称",
+  },
+  footer: {
+    copyright: "2026 World of ClaudeCraft",
+    githubLink: "https://github.com/levy-street/world-of-claudecraft",
+    githubLabel: "开源项目",
+    terms: "服务条款",
+    privacy: "隐私政策",
+    discordLabel: "加入Discord社区",
+  },
+  highscores: {
+    title: "高分排行榜",
+    desc: "追踪服务器中最伟大的英雄，并对比你的游戏进度。",
+  },
+  wiki: {
+    title: "游戏百科与指南",
+    desc: "探索世界的秘密、职业指南以及战术策略。",
+  },
+  news: {
+    title: "新闻与更新说明",
+    desc: "阅读最新补丁说明、游戏内事件和社区动态。",
+  },
+  download: {
+    title: "下载电脑版启动器",
+    desc: "获取独立启动器，以获得更优化的性能和全屏游戏体验。",
+  },
+  comingSoon: {
+    placeholder: "敬请期待...",
+    featureComingSoon: "该功能即将上线，敬请期待。",
+  },
+  mode: {
+    onlineTitle: "在线游玩",
+    onlineDesc: "登录服务器。你的角色数据保存在服务器上，与其他在线玩家共享同一个世界。",
+    onlineAria: "在线游玩：登录持久共享的联机世界",
+    offlineTitle: "单机试玩",
+    offlineDesc: "在浏览器中直接开启单人世界。所有数据均不保存：非常适合快速试玩或功能测试。",
+    offlineAria: "单机试玩：直接开始本地单人游戏会话",
+    tipTitle: "提示：",
+    tipText: "为了获得最流畅的游玩体验，请在此网站上关闭广告拦截插件。根据社区反馈，部分拦截器可能会导致游戏卡顿。",
+  },
+  auth: {
+    enterRealm: "进入服务器",
+    username: "用户名",
+    usernameError: "请输入用户名。",
+    usernamePlaceholder: "请输入用户名",
+    password: "密码",
+    passwordError: "请输入密码。",
+    passwordPlaceholder: "请输入密码",
+    showPassword: "显示密码",
+    hidePassword: "隐藏密码",
+    logIn: "登录",
+    createAccount: "创建账号",
+    back: "返回",
+    realmList: "服务器列表",
+    loadingRealms: "正在加载服务器...",
+    changeRealm: "切换服务器",
+    characters: "角色列表：",
+    createCharacter: "创建角色",
+    characterName: "角色名称",
+    characterNamePlaceholder: "输入角色名称",
+    enterWorld: "进入世界",
+    offlineCharacter: "单机角色",
+    create: "创建",
+  },
+  classes: {
+    warrior: "战士",
+    paladin: "圣骑士",
+    hunter: "猎人",
+    rogue: "潜行者",
+    priest: "牧师",
+    shaman: "萨满祭司",
+    mage: "法师",
+    warlock: "术士",
+    druid: "德鲁伊",
+    warriorAria: "战士职业",
+    paladinAria: "圣骑士职业",
+    hunterAria: "猎人职业",
+    rogueAria: "潜行者职业",
+    priestAria: "牧师职业",
+    shamanAria: "萨满祭司职业",
+    mageAria: "法师职业",
+    warlockAria: "术士职业",
+    druidAria: "德鲁伊职业",
+  },
+  controls: {
+    title: "操作指南",
+    movement: "移动操作",
+    moveTurn: "移动/转向",
+    strafe: "向左/向右平移",
+    jump: "跳跃",
+    autorun: "切换自动奔跑",
+    combat: "战斗与互动",
+    target: "锁定敌对目标",
+    spells: "施放法术",
+    interact: "互动/拾取",
+    nameplates: "显示姓名板",
+    camera: "视角与鼠标",
+    rightDrag: "右键拖拽",
+    leftDrag: "左键拖拽",
+    mouseWheel: "鼠标滚轮",
+    mouselook: "鼠标视口",
+    orbit: "环绕相机",
+    zoom: "缩放",
+    interfaces: "界面窗口",
+    charPane: "角色面板",
+    spellbook: "法术书",
+    questLog: "任务日志",
+    worldMap: "世界地图",
+    bags: "背包",
+    emoteWheel: "按住表情轮盘",
+    friends: "好友与公会",
+    chat: "打开聊天框",
+  },
+  ...shellStrings.zh_CN,
+  ...hudStrings.zh_CN,
+  ...abilityStrings.zh_CN,
+  ...questStrings.zh_CN,
+  ...itemStrings.zh_CN,
+  ...classAbilityNames.zh_CN,
+  ...itemNames.zh_CN,
+  ...worldNames.zh_CN,
+  ...mergeStrings.zh_CN,
+  entities: {
+    ...itemNames.zh_CN.entities,
+    ...worldNames.zh_CN.entities,
+    abilities: { ...itemNames.zh_CN.entities.abilities, ...mergeExtra.zh_CN.abilities },
+    items: { ...itemNames.zh_CN.entities.items, ...mergeEntities.zh_CN.items, ...mergeExtra.zh_CN.items },
+    mobs: { ...worldNames.zh_CN.entities.mobs, ...mergeEntities.zh_CN.mobs, ...mergeExtra.zh_CN.mobs },
+    npcs: { ...worldNames.zh_CN.entities.npcs, ...mergeExtra.zh_CN.npcs },
+    quests: { ...worldNames.zh_CN.entities.quests, ...mergeEntities.zh_CN.quests, ...mergeExtra.zh_CN.quests },
+    dungeons: { ...worldNames.zh_CN.entities.dungeons, ...mergeExtra.zh_CN.dungeons },
+  },
+};
