+++ src/ui/i18n.locales/zh_TW.ts
@@ -1,0 +1,169 @@
+import {
+  en,
+  shellStrings,
+  hudStrings,
+  abilityStrings,
+  questStrings,
+  itemStrings,
+  classAbilityNames,
+  itemNames,
+  mergeStrings,
+  mergeEntities,
+  mergeExtra,
+  gameStringsZhTW,
+} from '../i18n.en';
+import { worldEntityText as worldNames } from '../world_entity_i18n';
+
+export const zh_TW: typeof en = {
+  meta: { builtOn: "建置於 {date}" },
+  realmTypes: { normal: "一般", pvp: "PvP", rp: "RP", rpPvp: "RP-PvP" },
+  game: gameStringsZhTW,
+  nav: {
+    home: "首頁",
+    play: "開始遊戲",
+    stats: "數據統計",
+    about: "關於",
+    highscores: "排行榜",
+    wiki: "百科指南",
+    news: "最新消息",
+    download: "下載客戶端",
+    loginRegister: "登入/註冊",
+    donate: "捐贈",
+  },
+  stats: {
+    title: "伺服器狀態",
+    accountsCreated: "已建立帳號",
+    playersOnline: "線上玩家數",
+    realmName: "伺服器名稱",
+  },
+  footer: {
+    copyright: "2026 World of ClaudeCraft",
+    githubLink: "https://github.com/levy-street/world-of-claudecraft",
+    githubLabel: "開源專案",
+    terms: "服務條款",
+    privacy: "隱私權政策",
+    discordLabel: "加入Discord社區",
+  },
+  highscores: {
+    title: "高分排行榜",
+    desc: "追蹤伺服器中最偉大的英雄，並對比你的遊戲進度。",
+  },
+  wiki: {
+    title: "遊戲百科與指南",
+    desc: "探索世界的秘密、職業指南以及戰術策略。",
+  },
+  news: {
+    title: "新聞與更新說明",
+    desc: "閱讀最新修補程式說明、遊戲內事件和社群動態。",
+  },
+  download: {
+    title: "下載電腦版啟動器",
+    desc: "獲取獨立啟動器，以獲得更優化的效能和全螢幕遊戲體驗。",
+  },
+  comingSoon: {
+    placeholder: "敬請期待...",
+    featureComingSoon: "該功能即將上線，敬請期待。",
+  },
+  mode: {
+    onlineTitle: "線上遊玩",
+    onlineDesc: "登入伺服器。您的角色數據保存在伺服器上，與其他線上玩家共享同一個世界。",
+    onlineAria: "線上遊玩：登入持久共享的聯機世界",
+    offlineTitle: "單機試玩",
+    offlineDesc: "在瀏覽器中直接開啟單人世界。所有數據均不保存：非常適合快速試玩或功能測試。",
+    offlineAria: "單機試玩：直接開始本地單人遊戲會話",
+    tipTitle: "提示：",
+    tipText: "為了獲得最流暢的遊玩體驗，請在此網站上關閉廣告攔截外掛。根據社群回饋，部分攔截器可能會導致遊戲延遲。",
+  },
+  auth: {
+    enterRealm: "進入伺服器",
+    username: "使用者名稱",
+    usernameError: "請輸入使用者名稱。",
+    usernamePlaceholder: "請輸入使用者名稱",
+    password: "密碼",
+    passwordError: "請輸入密碼。",
+    passwordPlaceholder: "請輸入密碼",
+    showPassword: "顯示密碼",
+    hidePassword: "隱藏密碼",
+    logIn: "登入",
+    createAccount: "建立帳號",
+    back: "返回",
+    realmList: "伺服器列表",
+    loadingRealms: "正在載入伺服器...",
+    changeRealm: "切換伺服器",
+    characters: "角色列表：",
+    createCharacter: "建立角色",
+    characterName: "角色名稱",
+    characterNamePlaceholder: "輸入角色名稱",
+    enterWorld: "進入世界",
+    offlineCharacter: "單機角色",
+    create: "建立",
+  },
+  classes: {
+    warrior: "戰士",
+    paladin: "聖騎士",
+    hunter: "獵人",
+    rogue: "潛行者",
+    priest: "牧師",
+    shaman: "薩滿",
+    mage: "法師",
+    warlock: "術士",
+    druid: "德魯伊",
+    warriorAria: "戰士職業",
+    paladinAria: "聖騎士職業",
+    hunterAria: "獵人職業",
+    rogueAria: "潛行者職業",
+    priestAria: "牧師職業",
+    shamanAria: "薩滿職業",
+    mageAria: "法師職業",
+    warlockAria: "術士職業",
+    druidAria: "德魯伊職業",
+  },
+  controls: {
+    title: "操作指南",
+    movement: "移動操作",
+    moveTurn: "移動/轉向",
+    strafe: "向左/向右平移",
+    jump: "跳躍",
+    autorun: "切換自動奔跑",
+    combat: "戰鬥與互動",
+    target: "鎖定敵對目標",
+    spells: "施放法術",
+    interact: "互動/拾取",
+    nameplates: "顯示姓名板",
+    camera: "視角與滑鼠",
+    rightDrag: "右鍵拖曳",
+    leftDrag: "左鍵拖曳",
+    mouseWheel: "滑鼠滾輪",
+    mouselook: "滑鼠視角",
+    orbit: "環繞鏡頭",
+    zoom: "縮放",
+    interfaces: "介面視窗",
+    charPane: "角色面板",
+    spellbook: "法術書",
+    questLog: "任務日誌",
+    worldMap: "世界地圖",
+    bags: "背包背包欄",
+    emoteWheel: "按住表情輪盤",
+    friends: "好友與公會",
+    chat: "開啟對話框",
+  },
+  ...shellStrings.zh_TW,
+  ...hudStrings.zh_TW,
+  ...abilityStrings.zh_TW,
+  ...questStrings.zh_TW,
+  ...itemStrings.zh_TW,
+  ...classAbilityNames.zh_TW,
+  ...itemNames.zh_TW,
+  ...worldNames.zh_TW,
+  ...mergeStrings.zh_TW,
+  entities: {
+    ...itemNames.zh_TW.entities,
+    ...worldNames.zh_TW.entities,
+    abilities: { ...itemNames.zh_TW.entities.abilities, ...mergeExtra.zh_TW.abilities },
+    items: { ...itemNames.zh_TW.entities.items, ...mergeEntities.zh_TW.items, ...mergeExtra.zh_TW.items },
+    mobs: { ...worldNames.zh_TW.entities.mobs, ...mergeEntities.zh_TW.mobs, ...mergeExtra.zh_TW.mobs },
+    npcs: { ...worldNames.zh_TW.entities.npcs, ...mergeExtra.zh_TW.npcs },
+    quests: { ...worldNames.zh_TW.entities.quests, ...mergeEntities.zh_TW.quests, ...mergeExtra.zh_TW.quests },
+    dungeons: { ...worldNames.zh_TW.entities.dungeons, ...mergeExtra.zh_TW.dungeons },
+  },
+};
