+++ src/ui/i18n.ts
@@ -1,2040 +1,24 @@
-import {
-  en,
-  shellStrings,
-  hudStrings,
-  abilityStrings,
-  questStrings,
-  itemStrings,
-  classAbilityNames,
-  itemNames,
-  mergeStrings,
-  mergeEntities,
-  mergeExtra,
-  gameStringsEs,
-  gameStringsEsES,
-  gameStringsFrFR,
-  gameStringsFrCA,
-  gameStringsEnCA,
-  gameStringsItIT,
-  gameStringsDeDE,
-  gameStringsZhCN,
-  gameStringsZhTW,
-  gameStringsKoKR,
-  gameStringsJaJP,
-  gameStringsPtBR,
-  gameStringsRuRU,
-} from './i18n.en';
-import { worldEntityText as worldNames } from './world_entity_i18n';
+import { en } from './i18n.en';
+import { es } from './i18n.locales/es';
+import { es_ES } from './i18n.locales/es_ES';
+import { fr_FR } from './i18n.locales/fr_FR';
+import { fr_CA } from './i18n.locales/fr_CA';
+import { en_CA } from './i18n.locales/en_CA';
+import { it_IT } from './i18n.locales/it_IT';
+import { de_DE } from './i18n.locales/de_DE';
+import { zh_CN } from './i18n.locales/zh_CN';
+import { zh_TW } from './i18n.locales/zh_TW';
+import { ko_KR } from './i18n.locales/ko_KR';
+import { ja_JP } from './i18n.locales/ja_JP';
+import { pt_BR } from './i18n.locales/pt_BR';
+import { ru_RU } from './i18n.locales/ru_RU';
 import type { Leaves, TranslationKey, InterpolationValue, InterpolationValues, DeepPartial } from './i18n.en';
 
-// Re-export the authoritative en base, gameStrings, and the type machinery so
-// importers of './i18n' keep an unchanged public surface. The per-locale objects
-// are exported inline below.
-export { en };
+// Re-export the authoritative en base, every locale object, gameStrings, and the
+// type machinery so importers of './i18n' keep an unchanged public surface.
+export { en, es, es_ES, fr_FR, fr_CA, en_CA, it_IT, de_DE, zh_CN, zh_TW, ko_KR, ja_JP, pt_BR, ru_RU };
 export { gameStrings } from './i18n.en';
 export type { Leaves, TranslationKey, InterpolationValue, InterpolationValues, DeepPartial };
-
-export const es: typeof en = {
-  meta: { builtOn: "Compilado el {date}" },
-  realmTypes: { normal: "Normal", pvp: "JcJ", rp: "JdR", rpPvp: "JdR-JcJ" },
-  game: gameStringsEs,
-  nav: {
-    home: "Inicio",
-    play: "Jugar",
-    stats: "Estadísticas",
-    about: "Acerca de",
-    highscores: "Clasificaciones",
-    wiki: "Wiki",
-    news: "Noticias",
-    download: "Descargar",
-    loginRegister: "Iniciar Sesión/Registrarse",
-    donate: "Donar",
-  },
-  stats: {
-    title: "Estado del Reino",
-    accountsCreated: "Cuentas Creadas",
-    playersOnline: "Jugadores en Línea",
-    realmName: "Nombre del Reino",
-  },
-  footer: {
-    copyright: "2026 World of ClaudeCraft",
-    githubLink: "https://github.com/levy-street/world-of-claudecraft",
-    githubLabel: "Proyecto de Código Abierto",
-    terms: "Términos de Servicio",
-    privacy: "Política de Privacidad",
-    discordLabel: "Únete al Discord",
-  },
-  highscores: {
-    title: "Clasificaciones de Puntuación",
-    desc: "Sigue a los campeones más grandes del reino y compara tu progreso.",
-  },
-  wiki: {
-    title: "Wiki y Guía del Juego",
-    desc: "Descubre los secretos del reino, guías de clase y estrategias.",
-  },
-  news: {
-    title: "Noticias y Actualizaciones",
-    desc: "Lee las últimas notas del parche, eventos y noticias de la comunidad.",
-  },
-  download: {
-    title: "Descargar Lanzador de Escritorio",
-    desc: "Consigue el lanzador independiente para un rendimiento optimizado y juego a pantalla completa.",
-  },
-  comingSoon: {
-    placeholder: "Próximamente...",
-    featureComingSoon: "Esta característica llegará pronto al reino.",
-  },
-  mode: {
-    onlineTitle: "Jugar en Línea",
-    onlineDesc: "Inicia sesión en el reino. Tus personajes viven en el servidor y compartes el mundo con todos los demás que estén conectados.",
-    onlineAria: "Jugar en Línea: inicia sesión en el reino compartido persistente",
-    offlineTitle: "Jugar en Solitario",
-    offlineDesc: "Mundo instantáneo en solitario en tu navegador. Nada se guarda: perfecto para una pelea rápida o pruebas.",
-    offlineAria: "Jugar en Solitario: inicia una sesión local instantánea de un jugador",
-    tipTitle: "CONSEJO:",
-    tipText: "Para la mejor experiencia, desactiva las extensiones de bloqueo de publicidad en este sitio. Reportes de la comunidad encontraron que algunos bloqueadores pueden causar retrasos.",
-  },
-  auth: {
-    enterRealm: "Entrar al Reino",
-    username: "Usuario",
-    usernameError: "Por favor, introduce tu nombre de usuario.",
-    usernamePlaceholder: "Introduce tu usuario",
-    password: "Contraseña",
-    passwordError: "Por favor, introduce tu contraseña.",
-    passwordPlaceholder: "Introduce tu contraseña",
-    showPassword: "Mostrar contraseña",
-    hidePassword: "Ocultar contraseña",
-    logIn: "Iniciar Sesión",
-    createAccount: "Crear Cuenta",
-    back: "Atrás",
-    realmList: "Lista de Reinos",
-    loadingRealms: "Cargando reinos...",
-    changeRealm: "Cambiar de Reino",
-    characters: "Personajes:",
-    createCharacter: "Crear Personaje",
-    characterName: "Nombre del Personaje",
-    characterNamePlaceholder: "Nombre del personaje",
-    enterWorld: "Entrar al Mundo",
-    offlineCharacter: "Personaje en Solitario",
-    create: "Crear",
-  },
-  classes: {
-    warrior: "Guerrero",
-    paladin: "Paladín",
-    hunter: "Cazador",
-    rogue: "Pícaro",
-    priest: "Sacerdote",
-    shaman: "Chamán",
-    mage: "Mago",
-    warlock: "Brujo",
-    druid: "Druida",
-    warriorAria: "Clase Guerrero",
-    paladinAria: "Clase Paladín",
-    hunterAria: "Clase Cazador",
-    rogueAria: "Clase Pícaro",
-    priestAria: "Clase Sacerdote",
-    shamanAria: "Clase Chamán",
-    mageAria: "Clase Mago",
-    warlockAria: "Clase Brujo",
-    druidAria: "Clase Druida",
-  },
-  controls: {
-    title: "Guía de Controles",
-    movement: "Movimiento",
-    moveTurn: "Moverse / Girar",
-    strafe: "Desplazarse Izquierda/Derecha",
-    jump: "Saltar",
-    autorun: "Correr Automáticamente",
-    combat: "Combate e Interacción",
-    target: "Marcar Enemigo",
-    spells: "Lanzar Hechizos",
-    interact: "Interactuar / Despojar",
-    nameplates: "Mostrar Nombres",
-    camera: "Cámara y Ratón",
-    rightDrag: "Arrastrar Clic Derecho",
-    leftDrag: "Arrastrar Clic Izquierdo",
-    mouseWheel: "Rueda del Ratón",
-    mouselook: "Mirar con Ratón",
-    orbit: "Rotar Cámara",
-    zoom: "Zoom",
-    interfaces: "Interfaces",
-    charPane: "Panel de Personaje",
-    spellbook: "Libro de Hechizos",
-    questLog: "Diario de Misiones",
-    worldMap: "Mapa del Mundo",
-    bags: "Inventario de Bolsas",
-    emoteWheel: "Mantener rueda de gestos",
-    friends: "Amigos y Hermandad",
-    chat: "Abrir Chat",
-  },
-  ...shellStrings.es,
-  ...hudStrings.es,
-  ...abilityStrings.es,
-  ...questStrings.es,
-  ...itemStrings.es,
-  ...classAbilityNames.es,
-  ...itemNames.es,
-  ...worldNames.es,
-  ...mergeStrings.es,
-  entities: {
-    ...itemNames.es.entities,
-    ...worldNames.es.entities,
-    abilities: { ...itemNames.es.entities.abilities, ...mergeExtra.es.abilities },
-    items: { ...itemNames.es.entities.items, ...mergeEntities.es.items, ...mergeExtra.es.items },
-    mobs: { ...worldNames.es.entities.mobs, ...mergeEntities.es.mobs, ...mergeExtra.es.mobs },
-    npcs: { ...worldNames.es.entities.npcs, ...mergeExtra.es.npcs },
-    quests: { ...worldNames.es.entities.quests, ...mergeEntities.es.quests, ...mergeExtra.es.quests },
-    dungeons: { ...worldNames.es.entities.dungeons, ...mergeExtra.es.dungeons },
-  },
-};
-
-export const es_ES: typeof en = {
-  meta: { builtOn: "Compilado el {date}" },
-  realmTypes: { normal: "Normal", pvp: "JcJ", rp: "JdR", rpPvp: "JdR-JcJ" },
-  game: gameStringsEsES,
-  nav: {
-    home: "Inicio",
-    play: "Jugar",
-    stats: "Estadísticas",
-    about: "Acerca de",
-    highscores: "Clasificaciones",
-    wiki: "Wiki",
-    news: "Noticias",
-    download: "Descargar",
-    loginRegister: "Iniciar sesión/Registrarse",
-    donate: "Donar",
-  },
-  stats: {
-    title: "Estado del Reino",
-    accountsCreated: "Cuentas creadas",
-    playersOnline: "Jugadores en línea",
-    realmName: "Nombre del reino",
-  },
-  footer: {
-    copyright: "2026 World of ClaudeCraft",
-    githubLink: "https://github.com/levy-street/world-of-claudecraft",
-    githubLabel: "Proyecto de código abierto",
-    terms: "Términos de servicio",
-    privacy: "Política de privacidad",
-    discordLabel: "Únete al Discord",
-  },
-  highscores: {
-    title: "Tabla de clasificaciones",
-    desc: "Sigue a los campeones más grandes del reino y compara tu progreso.",
-  },
-  wiki: {
-    title: "Wiki y guía del juego",
-    desc: "Descubre los secretos del reino, guías de clase y estrategias.",
-  },
-  news: {
-    title: "Noticias y actualizaciones",
-    desc: "Lee las últimas notas del parche, eventos y noticias de la comunidad.",
-  },
-  download: {
-    title: "Descargar lanzador de escritorio",
-    desc: "Consigue el lanzador independiente para un rendimiento optimizado y juego a pantalla completa.",
-  },
-  comingSoon: {
-    placeholder: "Próximamente...",
-    featureComingSoon: "Esta característica llegará pronto al reino.",
-  },
-  mode: {
-    onlineTitle: "Jugar en línea",
-    onlineDesc: "Inicia sesión en el reino. Tus personajes viven en el servidor y compartes el mundo con todos los demás que estén conectados.",
-    onlineAria: "Jugar en línea: inicia sesión en el reino compartido persistente",
-    offlineTitle: "Jugar en solitario",
-    offlineDesc: "Mundo instantáneo en solitario en tu navegador. Nada se guarda: perfecto para una pelea rápida o pruebas.",
-    offlineAria: "Jugar en solitario: inicia una sesión local instantánea de un jugador",
-    tipTitle: "CONSEJO:",
-    tipText: "Para la mejor experiencia, desactiva las extensiones de bloqueo de publicidad en este sitio. Reportes de la comunidad encontraron que algunos bloqueadores pueden causar retrasos.",
-  },
-  auth: {
-    enterRealm: "Entrar al reino",
-    username: "Usuario",
-    usernameError: "Por favor, introduce tu nombre de usuario.",
-    usernamePlaceholder: "Introduce tu usuario",
-    password: "Contraseña",
-    passwordError: "Por favor, introduce tu contraseña.",
-    passwordPlaceholder: "Introduce tu contraseña",
-    showPassword: "Mostrar contraseña",
-    hidePassword: "Ocultar contraseña",
-    logIn: "Iniciar sesión",
-    createAccount: "Crear cuenta",
-    back: "Atrás",
-    realmList: "Lista de reinos",
-    loadingRealms: "Cargando reinos...",
-    changeRealm: "Cambiar de reino",
-    characters: "Personajes:",
-    createCharacter: "Crear personaje",
-    characterName: "Nombre del personaje",
-    characterNamePlaceholder: "Nombre del personaje",
-    enterWorld: "Entrar al mundo",
-    offlineCharacter: "Personaje en solitario",
-    create: "Crear",
-  },
-  classes: {
-    warrior: "Guerrero",
-    paladin: "Paladín",
-    hunter: "Cazador",
-    rogue: "Pícaro",
-    priest: "Sacerdote",
-    shaman: "Chamán",
-    mage: "Mago",
-    warlock: "Brujo",
-    druid: "Druida",
-    warriorAria: "Clase Guerrero",
-    paladinAria: "Clase Paladín",
-    hunterAria: "Clase Cazador",
-    rogueAria: "Clase Pícaro",
-    priestAria: "Clase Sacerdote",
-    shamanAria: "Clase Chamán",
-    mageAria: "Clase Mago",
-    warlockAria: "Clase Brujo",
-    druidAria: "Clase Druida",
-  },
-  controls: {
-    title: "Guía de controles",
-    movement: "Movimiento",
-    moveTurn: "Moverse/Girar",
-    strafe: "Desplazarse Izquierda/Derecha",
-    jump: "Saltar",
-    autorun: "Correr automáticamente",
-    combat: "Combate e interacción",
-    target: "Marcar enemigo",
-    spells: "Lanzar hechizos",
-    interact: "Interactuar/Despojar",
-    nameplates: "Mostrar nombres",
-    camera: "Cámara y ratón",
-    rightDrag: "Arrastrar clic derecho",
-    leftDrag: "Arrastrar clic izquierdo",
-    mouseWheel: "Rueda del ratón",
-    mouselook: "Mirar con ratón",
-    orbit: "Rotar cámara",
-    zoom: "Zoom",
-    interfaces: "Interfaces",
-    charPane: "Panel de personaje",
-    spellbook: "Libro de hechizos",
-    questLog: "Diario de misiones",
-    worldMap: "Mapa del mundo",
-    bags: "Inventario de bolsas",
-    emoteWheel: "Mantener rueda de gestos",
-    friends: "Amigos y hermandad",
-    chat: "Abrir chat",
-  },
-  ...shellStrings.es_ES,
-  ...hudStrings.es_ES,
-  ...abilityStrings.es_ES,
-  ...questStrings.es_ES,
-  ...itemStrings.es_ES,
-  ...classAbilityNames.es_ES,
-  ...itemNames.es_ES,
-  ...worldNames.es_ES,
-  ...mergeStrings.es_ES,
-  entities: {
-    ...itemNames.es_ES.entities,
-    ...worldNames.es_ES.entities,
-    abilities: { ...itemNames.es_ES.entities.abilities, ...mergeExtra.es_ES.abilities },
-    items: { ...itemNames.es_ES.entities.items, ...mergeEntities.es_ES.items, ...mergeExtra.es_ES.items },
-    mobs: { ...worldNames.es_ES.entities.mobs, ...mergeEntities.es_ES.mobs, ...mergeExtra.es_ES.mobs },
-    npcs: { ...worldNames.es_ES.entities.npcs, ...mergeExtra.es_ES.npcs },
-    quests: { ...worldNames.es_ES.entities.quests, ...mergeEntities.es_ES.quests, ...mergeExtra.es_ES.quests },
-    dungeons: { ...worldNames.es_ES.entities.dungeons, ...mergeExtra.es_ES.dungeons },
-  },
-};
-
-export const fr_FR: typeof en = {
-  meta: { builtOn: "Compilé le {date}" },
-  realmTypes: { normal: "Normal", pvp: "JcJ", rp: "JdR", rpPvp: "JdR-JcJ" },
-  game: gameStringsFrFR,
-  nav: {
-    home: "Accueil",
-    play: "Jouer",
-    stats: "Statistiques",
-    about: "À propos",
-    highscores: "Classements",
-    wiki: "Wiki",
-    news: "Actualités",
-    download: "Télécharger",
-    loginRegister: "Se connecter/S'inscrire",
-    donate: "Faire un don",
-  },
-  stats: {
-    title: "Statut du royaume",
-    accountsCreated: "Comptes créés",
-    playersOnline: "Joueurs en ligne",
-    realmName: "Nom du royaume",
-  },
-  footer: {
-    copyright: "2026 World of ClaudeCraft",
-    githubLink: "https://github.com/levy-street/world-of-claudecraft",
-    githubLabel: "Projet open source",
-    terms: "Conditions d'utilisation",
-    privacy: "Politique de confidentialité",
-    discordLabel: "Rejoindre le Discord",
-  },
-  highscores: {
-    title: "Tableau des classements",
-    desc: "Suivez les plus grands champions du royaume et comparez votre progression.",
-  },
-  wiki: {
-    title: "Wiki et guide du jeu",
-    desc: "Découvrez les secrets du royaume, les guides de classe et les stratégies.",
-  },
-  news: {
-    title: "Actualités et mises à jour",
-    desc: "Lisez les dernières notes de mise à jour, les événements et les nouvelles de la communauté.",
-  },
-  download: {
-    title: "Télécharger le lanceur de bureau",
-    desc: "Obtenez le lanceur autonome pour des performances optimisées et un jeu en plein écran.",
-  },
-  comingSoon: {
-    placeholder: "Bientôt disponible...",
-    featureComingSoon: "Cette fonctionnalité sera bientôt disponible sur le royaume.",
-  },
-  mode: {
-    onlineTitle: "Jouer en ligne",
-    onlineDesc: "Connectez-vous au royaume. Vos personnages vivent sur le serveur et vous partagez le monde avec tous les autres joueurs connectés.",
-    onlineAria: "Jouer en ligne: connectez-vous au royaume persistant partagé",
-    offlineTitle: "Jouer hors ligne",
-    offlineDesc: "Un monde solo instantané dans votre navigateur. Rien n'est sauvegardé: parfait pour un combat rapide ou des tests.",
-    offlineAria: "Jouer hors ligne: démarrer une session solo locale instantanée",
-    tipTitle: "CONSEIL:",
-    tipText: "Pour une expérience optimale, désactivez les extensions de blocage de publicité sur ce site. Certains bloqueurs peuvent causer des latences.",
-  },
-  auth: {
-    enterRealm: "Entrer dans le royaume",
-    username: "Nom d'utilisateur",
-    usernameError: "Veuillez saisir votre nom d'utilisateur.",
-    usernamePlaceholder: "Saisir le nom d'utilisateur",
-    password: "Mot de passe",
-    passwordError: "Veuillez saisir votre mot de passe.",
-    passwordPlaceholder: "Saisir le mot de passe",
-    showPassword: "Afficher le mot de passe",
-    hidePassword: "Masquer le mot de passe",
-    logIn: "Se connecter",
-    createAccount: "Créer un compte",
-    back: "Retour",
-    realmList: "Liste des royaumes",
-    loadingRealms: "Chargement des royaumes...",
-    changeRealm: "Changer de royaume",
-    characters: "Personnages:",
-    createCharacter: "Créer un personnage",
-    characterName: "Nom du personnage",
-    characterNamePlaceholder: "Nom du personnage",
-    enterWorld: "Entrer dans le monde",
-    offlineCharacter: "Personnage hors ligne",
-    create: "Créer",
-  },
-  classes: {
-    warrior: "Guerrier",
-    paladin: "Paladin",
-    hunter: "Chasseur",
-    rogue: "Voleur",
-    priest: "Prêtre",
-    shaman: "Chaman",
-    mage: "Mage",
-    warlock: "Démoniste",
-    druid: "Druide",
-    warriorAria: "Classe Guerrier",
-    paladinAria: "Classe Paladin",
-    hunterAria: "Classe Chasseur",
-    rogueAria: "Classe Voleur",
-    priestAria: "Classe Prêtre",
-    shamanAria: "Classe Chaman",
-    mageAria: "Classe Mage",
-    warlockAria: "Classe Démoniste",
-    druidAria: "Classe Druide",
-  },
-  controls: {
-    title: "Guide des contrôles",
-    movement: "Déplacement",
-    moveTurn: "Avancer/Tourner",
-    strafe: "Déplacement latéral Gauche/Droite",
-    jump: "Sauter",
-    autorun: "Activer la course automatique",
-    combat: "Combat et interaction",
-    target: "Cibler l'ennemi",
-    spells: "Lancer des sorts",
-    interact: "Interagir/Fouiller",
-    nameplates: "Afficher les plaques de nom",
-    camera: "Caméra et souris",
-    rightDrag: "Glisser-clic droit",
-    leftDrag: "Glisser-clic gauche",
-    mouseWheel: "Molette de la souris",
-    mouselook: "Vue à la souris",
-    orbit: "Caméra orbitale",
-    zoom: "Zoom",
-    interfaces: "Interfaces",
-    charPane: "Fiche de personnage",
-    spellbook: "Grimoire",
-    questLog: "Journal de quêtes",
-    worldMap: "Carte du monde",
-    bags: "Inventaire des sacs",
-    emoteWheel: "Maintenir la roue des emotes",
-    friends: "Amis et guilde",
-    chat: "Ouvrir le chat",
-  },
-  ...shellStrings.fr_FR,
-  ...hudStrings.fr_FR,
-  ...abilityStrings.fr_FR,
-  ...questStrings.fr_FR,
-  ...itemStrings.fr_FR,
-  ...classAbilityNames.fr_FR,
-  ...itemNames.fr_FR,
-  ...worldNames.fr_FR,
-  ...mergeStrings.fr_FR,
-  entities: {
-    ...itemNames.fr_FR.entities,
-    ...worldNames.fr_FR.entities,
-    abilities: { ...itemNames.fr_FR.entities.abilities, ...mergeExtra.fr_FR.abilities },
-    items: { ...itemNames.fr_FR.entities.items, ...mergeEntities.fr_FR.items, ...mergeExtra.fr_FR.items },
-    mobs: { ...worldNames.fr_FR.entities.mobs, ...mergeEntities.fr_FR.mobs, ...mergeExtra.fr_FR.mobs },
-    npcs: { ...worldNames.fr_FR.entities.npcs, ...mergeExtra.fr_FR.npcs },
-    quests: { ...worldNames.fr_FR.entities.quests, ...mergeEntities.fr_FR.quests, ...mergeExtra.fr_FR.quests },
-    dungeons: { ...worldNames.fr_FR.entities.dungeons, ...mergeExtra.fr_FR.dungeons },
-  },
-};
-
-export const fr_CA: typeof en = {
-  meta: { builtOn: "Compilé le {date}" },
-  realmTypes: { normal: "Normal", pvp: "JcJ", rp: "JdR", rpPvp: "JdR-JcJ" },
-  game: gameStringsFrCA,
-  nav: {
-    home: "Accueil",
-    play: "Jouer",
-    stats: "Statistiques",
-    about: "À propos",
-    highscores: "Classements",
-    wiki: "Wiki",
-    news: "Actualités",
-    download: "Télécharger",
-    loginRegister: "Se connecter/S'enregistrer",
-    donate: "Faire un don",
-  },
-  stats: {
-    title: "Statut du royaume",
-    accountsCreated: "Comptes créés",
-    playersOnline: "Joueurs en ligne",
-    realmName: "Nom du royaume",
-  },
-  footer: {
-    copyright: "2026 World of ClaudeCraft",
-    githubLink: "https://github.com/levy-street/world-of-claudecraft",
-    githubLabel: "Projet open source",
-    terms: "Conditions d'utilisation",
-    privacy: "Politique de confidentialité",
-    discordLabel: "Rejoindre le Discord",
-  },
-  highscores: {
-    title: "Tableau des classements",
-    desc: "Suivez les plus grands champions du royaume et comparez votre progression.",
-  },
-  wiki: {
-    title: "Wiki et guide du jeu",
-    desc: "Découvrez les secrets du royaume, les guides de classe et les stratégies.",
-  },
-  news: {
-    title: "Actualités et mises à jour",
-    desc: "Lisez les dernières notes de mise à jour, les événements et les nouvelles de la communauté.",
-  },
-  download: {
-    title: "Télécharger le lanceur de bureau",
-    desc: "Obtenez le lanceur autonome pour des performances optimisées et un jeu en plein écran.",
-  },
-  comingSoon: {
-    placeholder: "Bientôt disponible...",
-    featureComingSoon: "Cette fonctionnalité sera bientôt disponible sur le royaume.",
-  },
-  mode: {
-    onlineTitle: "Jouer en ligne",
-    onlineDesc: "Connectez-vous au royaume. Vos personnages vivent sur le serveur et vous partagez le monde avec tous les autres joueurs connectés.",
-    onlineAria: "Jouer en ligne: connectez-vous au royaume persistant partagé",
-    offlineTitle: "Jouer hors ligne",
-    offlineDesc: "Un monde solo instantané dans votre navigateur. Rien n'est sauvegardé: parfait pour un combat rapide ou des tests.",
-    offlineAria: "Jouer hors ligne: démarrer une session solo locale instantanée",
-    tipTitle: "CONSEIL:",
-    tipText: "Pour une expérience optimale, désactivez les extensions de blocage de publicité sur ce site. Certains bloqueurs peuvent causer des latences.",
-  },
-  auth: {
-    enterRealm: "Entrer dans le royaume",
-    username: "Nom d'utilisateur",
-    usernameError: "Veuillez saisir votre nom d'utilisateur.",
-    usernamePlaceholder: "Saisir le nom d'utilisateur",
-    password: "Mot de passe",
-    passwordError: "Veuillez saisir votre mot de passe.",
-    passwordPlaceholder: "Saisir le mot de passe",
-    showPassword: "Afficher le mot de passe",
-    hidePassword: "Masquer le mot de passe",
-    logIn: "Se connecter",
-    createAccount: "Créer un compte",
-    back: "Retour",
-    realmList: "Liste des royaumes",
-    loadingRealms: "Chargement des royaumes...",
-    changeRealm: "Changer de royaume",
-    characters: "Personnages:",
-    createCharacter: "Créer un personnage",
-    characterName: "Nom du personnage",
-    characterNamePlaceholder: "Nom du personnage",
-    enterWorld: "Entrer dans le monde",
-    offlineCharacter: "Personnage hors ligne",
-    create: "Créer",
-  },
-  classes: {
-    warrior: "Guerrier",
-    paladin: "Paladin",
-    hunter: "Chasseur",
-    rogue: "Voleur",
-    priest: "Prêtre",
-    shaman: "Chaman",
-    mage: "Mage",
-    warlock: "Démoniste",
-    druid: "Druide",
-    warriorAria: "Classe Guerrier",
-    paladinAria: "Classe Paladin",
-    hunterAria: "Classe Chasseur",
-    rogueAria: "Classe Voleur",
-    priestAria: "Classe Prêtre",
-    shamanAria: "Classe Chaman",
-    mageAria: "Classe Mage",
-    warlockAria: "Classe Démoniste",
-    druidAria: "Classe Druide",
-  },
-  controls: {
-    title: "Guide des contrôles",
-    movement: "Déplacement",
-    moveTurn: "Avancer/Tourner",
-    strafe: "Déplacement latéral Gauche/Droite",
-    jump: "Sauter",
-    autorun: "Activer la course automatique",
-    combat: "Combat et interaction",
-    target: "Cibler l'ennemi",
-    spells: "Lancer des sorts",
-    interact: "Interagir/Fouiller",
-    nameplates: "Afficher les plaques de nom",
-    camera: "Caméra et souris",
-    rightDrag: "Glisser-clic droit",
-    leftDrag: "Glisser-clic gauche",
-    mouseWheel: "Molette de la souris",
-    mouselook: "Vue à la souris",
-    orbit: "Caméra orbitale",
-    zoom: "Zoom",
-    interfaces: "Interfaces",
-    charPane: "Fiche de personnage",
-    spellbook: "Grimoire",
-    questLog: "Journal de quêtes",
-    worldMap: "Carte du monde",
-    bags: "Inventaire des sacs",
-    emoteWheel: "Maintenir la roue des emotes",
-    friends: "Amis et guilde",
-    chat: "Ouvrir le chat",
-  },
-  ...shellStrings.fr_CA,
-  ...hudStrings.fr_CA,
-  ...abilityStrings.fr_CA,
-  ...questStrings.fr_CA,
-  ...itemStrings.fr_CA,
-  ...classAbilityNames.fr_CA,
-  ...itemNames.fr_CA,
-  ...worldNames.fr_CA,
-  ...mergeStrings.fr_CA,
-  entities: {
-    ...itemNames.fr_CA.entities,
-    ...worldNames.fr_CA.entities,
-    abilities: { ...itemNames.fr_CA.entities.abilities, ...mergeExtra.fr_CA.abilities },
-    items: { ...itemNames.fr_CA.entities.items, ...mergeEntities.fr_CA.items, ...mergeExtra.fr_CA.items },
-    mobs: { ...worldNames.fr_CA.entities.mobs, ...mergeEntities.fr_CA.mobs, ...mergeExtra.fr_CA.mobs },
-    npcs: { ...worldNames.fr_CA.entities.npcs, ...mergeExtra.fr_CA.npcs },
-    quests: { ...worldNames.fr_CA.entities.quests, ...mergeEntities.fr_CA.quests, ...mergeExtra.fr_CA.quests },
-    dungeons: { ...worldNames.fr_CA.entities.dungeons, ...mergeExtra.fr_CA.dungeons },
-  },
-};
-
-export const en_CA: typeof en = {
-  meta: { builtOn: "Built {date}" },
-  realmTypes: { normal: "Normal", pvp: "PvP", rp: "RP", rpPvp: "RP-PvP" },
-  game: gameStringsEnCA,
-  nav: {
-    home: "Home",
-    play: "Play",
-    stats: "Statistics",
-    about: "About",
-    highscores: "High Scores",
-    wiki: "Wiki",
-    news: "News",
-    download: "Download",
-    loginRegister: "Login/Register",
-    donate: "Donate",
-  },
-  stats: {
-    title: "Realm Status",
-    accountsCreated: "Accounts Created",
-    playersOnline: "Players Online",
-    realmName: "Realm Name",
-  },
-  footer: {
-    copyright: "2026 World of ClaudeCraft",
-    githubLink: "https://github.com/levy-street/world-of-claudecraft",
-    githubLabel: "Open Source Project",
-    terms: "Terms of Service",
-    privacy: "Privacy Policy",
-    discordLabel: "Join the Discord",
-  },
-  highscores: {
-    title: "High Scores Leaderboard",
-    desc: "Track the realm's greatest champions and compare your progress.",
-  },
-  wiki: {
-    title: "Game Wiki & Guide",
-    desc: "Discover the secrets of the realm, class guides, and strategies.",
-  },
-  news: {
-    title: "News & Updates",
-    desc: "Read the latest patch notes, events, and community updates.",
-  },
-  download: {
-    title: "Download Desktop Launcher",
-    desc: "Get the standalone launcher for optimized performance and full-screen play.",
-  },
-  comingSoon: {
-    placeholder: "Coming Soon...",
-    featureComingSoon: "This feature is coming soon to the realm.",
-  },
-  mode: {
-    onlineTitle: "Play Online",
-    onlineDesc: "Log in to the realm. Your characters live on the server and you share the world with everyone else who's on.",
-    onlineAria: "Play Online: log in to the persistent shared realm",
-    offlineTitle: "Play Offline",
-    offlineDesc: "Instant single-player world in your browser. Nothing is saved: perfect for a quick brawl or testing.",
-    offlineAria: "Play Offline: start an instant local single-player session",
-    tipTitle: "TIP:",
-    tipText: "For the smoothest experience, turn off ad blocker extensions on this site. Community reports found some blockers can cause lag.",
-  },
-  auth: {
-    enterRealm: "Enter the Realm",
-    username: "Username",
-    usernameError: "Please enter your username.",
-    usernamePlaceholder: "Enter username",
-    password: "Password",
-    passwordError: "Please enter your password.",
-    passwordPlaceholder: "Enter password",
-    showPassword: "Show password",
-    hidePassword: "Hide password",
-    logIn: "Log In",
-    createAccount: "Create Account",
-    back: "Back",
-    realmList: "Realm List",
-    loadingRealms: "Loading realms...",
-    changeRealm: "Change Realm",
-    characters: "Characters:",
-    createCharacter: "Create Character",
-    characterName: "Character Name",
-    characterNamePlaceholder: "Character name",
-    enterWorld: "Enter World",
-    offlineCharacter: "Offline Character",
-    create: "Create",
-  },
-  classes: {
-    warrior: "Warrior",
-    paladin: "Paladin",
-    hunter: "Hunter",
-    rogue: "Rogue",
-    priest: "Priest",
-    shaman: "Shaman",
-    mage: "Mage",
-    warlock: "Warlock",
-    druid: "Druid",
-    warriorAria: "Warrior class",
-    paladinAria: "Paladin class",
-    hunterAria: "Hunter class",
-    rogueAria: "Rogue class",
-    priestAria: "Priest class",
-    shamanAria: "Shaman class",
-    mageAria: "Mage class",
-    warlockAria: "Warlock class",
-    druidAria: "Druid class",
-  },
-  controls: {
-    title: "Controls Guide",
-    movement: "Movement",
-    moveTurn: "Move / Turn",
-    strafe: "Strafe Left/Right",
-    jump: "Jump",
-    autorun: "Toggle Autorun",
-    combat: "Combat & Interaction",
-    target: "Target Enemy",
-    spells: "Cast Spells",
-    interact: "Interact / Loot",
-    nameplates: "Toggle Nameplates",
-    camera: "Camera & Mouse",
-    rightDrag: "Right-Drag",
-    leftDrag: "Left-Drag",
-    mouseWheel: "Mouse Wheel",
-    mouselook: "Mouselook",
-    orbit: "Orbit Camera",
-    zoom: "Zoom",
-    interfaces: "Interfaces",
-    charPane: "Character Pane",
-    spellbook: "Spellbook",
-    questLog: "Quest Log",
-    worldMap: "World Map",
-    bags: "Bags Inventory",
-    emoteWheel: "Hold Emote Wheel",
-    friends: "Friends & Guild",
-    chat: "Open Chat",
-  },
-  ...shellStrings.en_CA,
-  ...hudStrings.en_CA,
-  ...abilityStrings.en_CA,
-  ...questStrings.en_CA,
-  ...itemStrings.en_CA,
-  ...classAbilityNames.en_CA,
-  ...itemNames.en_CA,
-  ...worldNames.en_CA,
-  ...mergeStrings.en_CA,
-  entities: {
-    ...itemNames.en_CA.entities,
-    ...worldNames.en_CA.entities,
-    abilities: { ...itemNames.en_CA.entities.abilities, ...mergeExtra.en_CA.abilities },
-    items: { ...itemNames.en_CA.entities.items, ...mergeEntities.en_CA.items, ...mergeExtra.en_CA.items },
-    mobs: { ...worldNames.en_CA.entities.mobs, ...mergeEntities.en_CA.mobs, ...mergeExtra.en_CA.mobs },
-    npcs: { ...worldNames.en_CA.entities.npcs, ...mergeExtra.en_CA.npcs },
-    quests: { ...worldNames.en_CA.entities.quests, ...mergeEntities.en_CA.quests, ...mergeExtra.en_CA.quests },
-    dungeons: { ...worldNames.en_CA.entities.dungeons, ...mergeExtra.en_CA.dungeons },
-  },
-};
-
-export const it_IT: typeof en = {
-  meta: { builtOn: "Compilato il {date}" },
-  realmTypes: { normal: "Normale", pvp: "PvP", rp: "GdR", rpPvp: "GdR-PvP" },
-  game: gameStringsItIT,
-  nav: {
-    home: "Pagina principale",
-    play: "Gioca",
-    stats: "Statistiche",
-    about: "Informazioni",
-    highscores: "Classifiche",
-    wiki: "Wiki",
-    news: "Novità",
-    download: "Scarica",
-    loginRegister: "Accedi/Registrati",
-    donate: "Dona",
-  },
-  stats: {
-    title: "Stato del Reame",
-    accountsCreated: "Account creati",
-    playersOnline: "Giocatori online",
-    realmName: "Nome del reame",
-  },
-  footer: {
-    copyright: "2026 World of ClaudeCraft",
-    githubLink: "https://github.com/levy-street/world-of-claudecraft",
-    githubLabel: "Progetto open source",
-    terms: "Termini di servizio",
-    privacy: "Informativa sulla privacy",
-    discordLabel: "Entra nel Discord",
-  },
-  highscores: {
-    title: "Classifica dei punteggi più alti",
-    desc: "Segui i più grandi campioni del reame e confronta i tuoi progressi.",
-  },
-  wiki: {
-    title: "Wiki e guida del gioco",
-    desc: "Scopri i segreti del reame, le guide di classe e le strategie.",
-  },
-  news: {
-    title: "Notizie e aggiornamenti",
-    desc: "Leggi le ultime note sulla patch, gli eventi e gli aggiornamenti della community.",
-  },
-  download: {
-    title: "Scarica il launcher per desktop",
-    desc: "Ottieni il launcher autonomo per prestazioni ottimizzate e gioco a schermo intero.",
-  },
-  comingSoon: {
-    placeholder: "Prossimamente...",
-    featureComingSoon: "Questa funzione sarà presto disponibile nel reame.",
-  },
-  mode: {
-    onlineTitle: "Gioca online",
-    onlineDesc: "Accedi al reame. I tuoi personaggi vivono sul server e condividi il mondo con tutti gli altri giocatori connessi.",
-    onlineAria: "Gioca online: accedi al reame condiviso persistente",
-    offlineTitle: "Gioca offline",
-    offlineDesc: "Un mondo a giocatore singolo istantaneo nel tuo browser. Nulla viene salvato: perfetto per un combattimento rapido o per testare.",
-    offlineAria: "Gioca offline: avvia una sessione locale a giocatore singolo istantanea",
-    tipTitle: "SUGGERIMENTO:",
-    tipText: "Per un'esperienza ottimale, disattiva le estensioni di blocco della pubblicità su questo sito. Alcuni utenti segnalano che i blocchi possono causare rallentamenti.",
-  },
-  auth: {
-    enterRealm: "Entra nel reame",
-    username: "Nome utente",
-    usernameError: "Inserisci il tuo nome utente.",
-    usernamePlaceholder: "Inserisci il nome utente",
-    password: "Password",
-    passwordError: "Inserisci la tua password.",
-    passwordPlaceholder: "Inserisci la password",
-    showPassword: "Mostra password",
-    hidePassword: "Nascondi password",
-    logIn: "Accedi",
-    createAccount: "Crea account",
-    back: "Indietro",
-    realmList: "Lista dei reami",
-    loadingRealms: "Caricamento dei reami...",
-    changeRealm: "Cambia reame",
-    characters: "Personaggi:",
-    createCharacter: "Crea personaggio",
-    characterName: "Nome del personaggio",
-    characterNamePlaceholder: "Nome del personaggio",
-    enterWorld: "Entra nel mondo",
-    offlineCharacter: "Personaggio offline",
-    create: "Crea",
-  },
-  classes: {
-    warrior: "Guerriero",
-    paladin: "Paladino",
-    hunter: "Cacciatore",
-    rogue: "Ladro",
-    priest: "Sacerdote",
-    shaman: "Sciamano",
-    mage: "Mago",
-    warlock: "Stregone",
-    druid: "Druido",
-    warriorAria: "Classe Guerriero",
-    paladinAria: "Classe Paladino",
-    hunterAria: "Classe Cacciatore",
-    rogueAria: "Classe Ladro",
-    priestAria: "Classe Sacerdote",
-    shamanAria: "Classe Sciamano",
-    mageAria: "Classe Mago",
-    warlockAria: "Classe Stregone",
-    druidAria: "Classe Druido",
-  },
-  controls: {
-    title: "Guida ai controlli",
-    movement: "Movimento",
-    moveTurn: "Muoviti/Gira",
-    strafe: "Passi laterali Sinistra/Destra",
-    jump: "Salta",
-    autorun: "Attiva corsa automatica",
-    combat: "Combattimento e interazione",
-    target: "Seleziona nemico come bersaglio",
-    spells: "Lancia incantesimi",
-    interact: "Interagisci/Saccheggia",
-    nameplates: "Attiva targhette dei nomi",
-    camera: "Telecamera e mouse",
-    rightDrag: "Trascina con il tasto destro",
-    leftDrag: "Trascina con il tasto sinistro",
-    mouseWheel: "Rotella del mouse",
-    mouselook: "Visuale con il mouse",
-    orbit: "Orbita telecamera",
-    zoom: "Zoom",
-    interfaces: "Interfacce",
-    charPane: "Schermata personaggio",
-    spellbook: "Libro degli incantesimi",
-    questLog: "Registro delle missioni",
-    worldMap: "Mappa del mondo",
-    bags: "Inventario borse",
-    emoteWheel: "Tieni premuta la ruota emote",
-    friends: "Amici e gilda",
-    chat: "Apri chat",
-  },
-  ...shellStrings.it_IT,
-  ...hudStrings.it_IT,
-  ...abilityStrings.it_IT,
-  ...questStrings.it_IT,
-  ...itemStrings.it_IT,
-  ...classAbilityNames.it_IT,
-  ...itemNames.it_IT,
-  ...worldNames.it_IT,
-  ...mergeStrings.it_IT,
-  entities: {
-    ...itemNames.it_IT.entities,
-    ...worldNames.it_IT.entities,
-    abilities: { ...itemNames.it_IT.entities.abilities, ...mergeExtra.it_IT.abilities },
-    items: { ...itemNames.it_IT.entities.items, ...mergeEntities.it_IT.items, ...mergeExtra.it_IT.items },
-    mobs: { ...worldNames.it_IT.entities.mobs, ...mergeEntities.it_IT.mobs, ...mergeExtra.it_IT.mobs },
-    npcs: { ...worldNames.it_IT.entities.npcs, ...mergeExtra.it_IT.npcs },
-    quests: { ...worldNames.it_IT.entities.quests, ...mergeEntities.it_IT.quests, ...mergeExtra.it_IT.quests },
-    dungeons: { ...worldNames.it_IT.entities.dungeons, ...mergeExtra.it_IT.dungeons },
-  },
-};
-
-export const de_DE: typeof en = {
-  meta: { builtOn: "Erstellt am {date}" },
-  realmTypes: { normal: "Normal", pvp: "PvP", rp: "RP", rpPvp: "RP-PvP" },
-  game: gameStringsDeDE,
-  nav: {
-    home: "Startseite",
-    play: "Spielen",
-    stats: "Statistiken",
-    about: "Über uns",
-    highscores: "Bestenlisten",
-    wiki: "Wiki",
-    news: "Neuigkeiten",
-    download: "Herunterladen",
-    loginRegister: "Einloggen/Registrieren",
-    donate: "Spenden",
-  },
-  stats: {
-    title: "Realmstatus",
-    accountsCreated: "Erstellte Accounts",
-    playersOnline: "Spieler online",
-    realmName: "Realmname",
-  },
-  footer: {
-    copyright: "2026 World of ClaudeCraft",
-    githubLink: "https://github.com/levy-street/world-of-claudecraft",
-    githubLabel: "Open-Source-Projekt",
-    terms: "Nutzungsbedingungen",
-    privacy: "Datenschutzerklärung",
-    discordLabel: "Tritt dem Discord bei",
-  },
-  highscores: {
-    title: "Bestenliste",
-    desc: "Verfolge die größten Champions des Realms und vergleiche deinen Fortschritt.",
-  },
-  wiki: {
-    title: "Spiel-Wiki und Guides",
-    desc: "Entdecke die Geheimnisse des Realms, Klassenguides und Strategien.",
-  },
-  news: {
-    title: "Neuigkeiten und Updates",
-    desc: "Lies die neuesten Patchnotes, Events und Community-Updates.",
-  },
-  download: {
-    title: "Desktop-Launcher herunterladen",
-    desc: "Hol dir den eigenständigen Launcher für optimierte Leistung und Vollbildspiel.",
-  },
-  comingSoon: {
-    placeholder: "Demnächst...",
-    featureComingSoon: "Diese Funktion ist demnächst auf dem Realm verfügbar.",
-  },
-  mode: {
-    onlineTitle: "Online spielen",
-    onlineDesc: "Logge dich im Realm ein. Deine Charaktere befinden sich auf dem Server und du teilst dir die Welt mit allen anderen Spielern online.",
-    onlineAria: "Online spielen: Logge dich im persistenten, geteilten Realm ein",
-    offlineTitle: "Offline spielen",
-    offlineDesc: "Sofortige Einzelspielerwelt in deinem Browser. Es wird nichts gespeichert: perfekt für einen schnellen Kampf oder zum Testen.",
-    offlineAria: "Offline spielen: Starte eine sofortige lokale Einzelspielersitzung",
-    tipTitle: "TIPP:",
-    tipText: "Deaktiviere für das flüssigste Spielerlebnis Adblocker-Erweiterungen auf dieser Website. Community-Berichte zeigen, dass einige Blocker zu Verzögerungen führen können.",
-  },
-  auth: {
-    enterRealm: "Realm betreten",
-    username: "Benutzername",
-    usernameError: "Bitte gib deinen Benutzernamen ein.",
-    usernamePlaceholder: "Benutzername eingeben",
-    password: "Passwort",
-    passwordError: "Bitte gib dein Passwort ein.",
-    passwordPlaceholder: "Passwort eingeben",
-    showPassword: "Passwort anzeigen",
-    hidePassword: "Passwort ausblenden",
-    logIn: "Einloggen",
-    createAccount: "Account erstellen",
-    back: "Zurück",
-    realmList: "Realmliste",
-    loadingRealms: "Realms werden geladen...",
-    changeRealm: "Realm wechseln",
-    characters: "Charaktere:",
-    createCharacter: "Charakter erstellen",
-    characterName: "Charaktername",
-    characterNamePlaceholder: "Charaktername",
-    enterWorld: "Welt betreten",
-    offlineCharacter: "Offline-Charakter",
-    create: "Erstellen",
-  },
-  classes: {
-    warrior: "Krieger",
-    paladin: "Paladin",
-    hunter: "Jäger",
-    rogue: "Schurke",
-    priest: "Priester",
-    shaman: "Schamane",
-    mage: "Magier",
-    warlock: "Hexenmeister",
-    druid: "Druide",
-    warriorAria: "Krieger-Klasse",
-    paladinAria: "Paladin-Klasse",
-    hunterAria: "Jäger-Klasse",
-    rogueAria: "Schurke-Klasse",
-    priestAria: "Priester-Klasse",
-    shamanAria: "Schamane-Klasse",
-    mageAria: "Magier-Klasse",
-    warlockAria: "Hexenmeister-Klasse",
-    druidAria: "Druide-Klasse",
-  },
-  controls: {
-    title: "Steuerungsanleitung",
-    movement: "Bewegung",
-    moveTurn: "Bewegen/Drehen",
-    strafe: "Seitwärts bewegen Links/Rechts",
-    jump: "Springen",
-    autorun: "Automatisches Laufen umschalten",
-    combat: "Kampf und Interaktion",
-    target: "Feind anvisieren",
-    spells: "Zauber wirken",
-    interact: "Interagieren/Plündern",
-    nameplates: "Namensplaketten umschalten",
-    camera: "Kamera und Maus",
-    rightDrag: "Rechts ziehen",
-    leftDrag: "Links ziehen",
-    mouseWheel: "Mausrad",
-    mouselook: "Mausblick",
-    orbit: "Kamera umkreisen",
-    zoom: "Zoom",
-    interfaces: "Benutzeroberflächen",
-    charPane: "Charakterfenster",
-    spellbook: "Zauberbuch",
-    questLog: "Questlog",
-    worldMap: "Weltkarte",
-    bags: "Tascheninventar",
-    emoteWheel: "Emote-Rad gedrückt halten",
-    friends: "Freunde und Gilde",
-    chat: "Chat öffnen",
-  },
-  ...shellStrings.de_DE,
-  ...hudStrings.de_DE,
-  ...abilityStrings.de_DE,
-  ...questStrings.de_DE,
-  ...itemStrings.de_DE,
-  ...classAbilityNames.de_DE,
-  ...itemNames.de_DE,
-  ...worldNames.de_DE,
-  ...mergeStrings.de_DE,
-  entities: {
-    ...itemNames.de_DE.entities,
-    ...worldNames.de_DE.entities,
-    abilities: { ...itemNames.de_DE.entities.abilities, ...mergeExtra.de_DE.abilities },
-    items: { ...itemNames.de_DE.entities.items, ...mergeEntities.de_DE.items, ...mergeExtra.de_DE.items },
-    mobs: { ...worldNames.de_DE.entities.mobs, ...mergeEntities.de_DE.mobs, ...mergeExtra.de_DE.mobs },
-    npcs: { ...worldNames.de_DE.entities.npcs, ...mergeExtra.de_DE.npcs },
-    quests: { ...worldNames.de_DE.entities.quests, ...mergeEntities.de_DE.quests, ...mergeExtra.de_DE.quests },
-    dungeons: { ...worldNames.de_DE.entities.dungeons, ...mergeExtra.de_DE.dungeons },
-  },
-};
-
-export const zh_CN: typeof en = {
-  meta: { builtOn: "构建于 {date}" },
-  realmTypes: { normal: "普通", pvp: "PvP", rp: "RP", rpPvp: "RP-PvP" },
-  game: gameStringsZhCN,
-  nav: {
-    home: "首页",
-    play: "开始游戏",
-    stats: "数据统计",
-    about: "关于",
-    highscores: "排行榜",
-    wiki: "百科指南",
-    news: "新闻资讯",
-    download: "下载客户端",
-    loginRegister: "登录/注册",
-    donate: "捐赠",
-  },
-  stats: {
-    title: "服务器状态",
-    accountsCreated: "已创建账号",
-    playersOnline: "在线玩家数",
-    realmName: "服务器名称",
-  },
-  footer: {
-    copyright: "2026 World of ClaudeCraft",
-    githubLink: "https://github.com/levy-street/world-of-claudecraft",
-    githubLabel: "开源项目",
-    terms: "服务条款",
-    privacy: "隐私政策",
-    discordLabel: "加入Discord社区",
-  },
-  highscores: {
-    title: "高分排行榜",
-    desc: "追踪服务器中最伟大的英雄，并对比你的游戏进度。",
-  },
-  wiki: {
-    title: "游戏百科与指南",
-    desc: "探索世界的秘密、职业指南以及战术策略。",
-  },
-  news: {
-    title: "新闻与更新说明",
-    desc: "阅读最新补丁说明、游戏内事件和社区动态。",
-  },
-  download: {
-    title: "下载电脑版启动器",
-    desc: "获取独立启动器，以获得更优化的性能和全屏游戏体验。",
-  },
-  comingSoon: {
-    placeholder: "敬请期待...",
-    featureComingSoon: "该功能即将上线，敬请期待。",
-  },
-  mode: {
-    onlineTitle: "在线游玩",
-    onlineDesc: "登录服务器。你的角色数据保存在服务器上，与其他在线玩家共享同一个世界。",
-    onlineAria: "在线游玩：登录持久共享的联机世界",
-    offlineTitle: "单机试玩",
-    offlineDesc: "在浏览器中直接开启单人世界。所有数据均不保存：非常适合快速试玩或功能测试。",
-    offlineAria: "单机试玩：直接开始本地单人游戏会话",
-    tipTitle: "提示：",
-    tipText: "为了获得最流畅的游玩体验，请在此网站上关闭广告拦截插件。根据社区反馈，部分拦截器可能会导致游戏卡顿。",
-  },
-  auth: {
-    enterRealm: "进入服务器",
-    username: "用户名",
-    usernameError: "请输入用户名。",
-    usernamePlaceholder: "请输入用户名",
-    password: "密码",
-    passwordError: "请输入密码。",
-    passwordPlaceholder: "请输入密码",
-    showPassword: "显示密码",
-    hidePassword: "隐藏密码",
-    logIn: "登录",
-    createAccount: "创建账号",
-    back: "返回",
-    realmList: "服务器列表",
-    loadingRealms: "正在加载服务器...",
-    changeRealm: "切换服务器",
-    characters: "角色列表：",
-    createCharacter: "创建角色",
-    characterName: "角色名称",
-    characterNamePlaceholder: "输入角色名称",
-    enterWorld: "进入世界",
-    offlineCharacter: "单机角色",
-    create: "创建",
-  },
-  classes: {
-    warrior: "战士",
-    paladin: "圣骑士",
-    hunter: "猎人",
-    rogue: "潜行者",
-    priest: "牧师",
-    shaman: "萨满祭司",
-    mage: "法师",
-    warlock: "术士",
-    druid: "德鲁伊",
-    warriorAria: "战士职业",
-    paladinAria: "圣骑士职业",
-    hunterAria: "猎人职业",
-    rogueAria: "潜行者职业",
-    priestAria: "牧师职业",
-    shamanAria: "萨满祭司职业",
-    mageAria: "法师职业",
-    warlockAria: "术士职业",
-    druidAria: "德鲁伊职业",
-  },
-  controls: {
-    title: "操作指南",
-    movement: "移动操作",
-    moveTurn: "移动/转向",
-    strafe: "向左/向右平移",
-    jump: "跳跃",
-    autorun: "切换自动奔跑",
-    combat: "战斗与互动",
-    target: "锁定敌对目标",
-    spells: "施放法术",
-    interact: "互动/拾取",
-    nameplates: "显示姓名板",
-    camera: "视角与鼠标",
-    rightDrag: "右键拖拽",
-    leftDrag: "左键拖拽",
-    mouseWheel: "鼠标滚轮",
-    mouselook: "鼠标视口",
-    orbit: "环绕相机",
-    zoom: "缩放",
-    interfaces: "界面窗口",
-    charPane: "角色面板",
-    spellbook: "法术书",
-    questLog: "任务日志",
-    worldMap: "世界地图",
-    bags: "背包",
-    emoteWheel: "按住表情轮盘",
-    friends: "好友与公会",
-    chat: "打开聊天框",
-  },
-  ...shellStrings.zh_CN,
-  ...hudStrings.zh_CN,
-  ...abilityStrings.zh_CN,
-  ...questStrings.zh_CN,
-  ...itemStrings.zh_CN,
-  ...classAbilityNames.zh_CN,
-  ...itemNames.zh_CN,
-  ...worldNames.zh_CN,
-  ...mergeStrings.zh_CN,
-  entities: {
-    ...itemNames.zh_CN.entities,
-    ...worldNames.zh_CN.entities,
-    abilities: { ...itemNames.zh_CN.entities.abilities, ...mergeExtra.zh_CN.abilities },
-    items: { ...itemNames.zh_CN.entities.items, ...mergeEntities.zh_CN.items, ...mergeExtra.zh_CN.items },
-    mobs: { ...worldNames.zh_CN.entities.mobs, ...mergeEntities.zh_CN.mobs, ...mergeExtra.zh_CN.mobs },
-    npcs: { ...worldNames.zh_CN.entities.npcs, ...mergeExtra.zh_CN.npcs },
-    quests: { ...worldNames.zh_CN.entities.quests, ...mergeEntities.zh_CN.quests, ...mergeExtra.zh_CN.quests },
-    dungeons: { ...worldNames.zh_CN.entities.dungeons, ...mergeExtra.zh_CN.dungeons },
-  },
-};
-
-export const zh_TW: typeof en = {
-  meta: { builtOn: "建置於 {date}" },
-  realmTypes: { normal: "一般", pvp: "PvP", rp: "RP", rpPvp: "RP-PvP" },
-  game: gameStringsZhTW,
-  nav: {
-    home: "首頁",
-    play: "開始遊戲",
-    stats: "數據統計",
-    about: "關於",
-    highscores: "排行榜",
-    wiki: "百科指南",
-    news: "最新消息",
-    download: "下載客戶端",
-    loginRegister: "登入/註冊",
-    donate: "捐贈",
-  },
-  stats: {
-    title: "伺服器狀態",
-    accountsCreated: "已建立帳號",
-    playersOnline: "線上玩家數",
-    realmName: "伺服器名稱",
-  },
-  footer: {
-    copyright: "2026 World of ClaudeCraft",
-    githubLink: "https://github.com/levy-street/world-of-claudecraft",
-    githubLabel: "開源專案",
-    terms: "服務條款",
-    privacy: "隱私權政策",
-    discordLabel: "加入Discord社區",
-  },
-  highscores: {
-    title: "高分排行榜",
-    desc: "追蹤伺服器中最偉大的英雄，並對比你的遊戲進度。",
-  },
-  wiki: {
-    title: "遊戲百科與指南",
-    desc: "探索世界的秘密、職業指南以及戰術策略。",
-  },
-  news: {
-    title: "新聞與更新說明",
-    desc: "閱讀最新修補程式說明、遊戲內事件和社群動態。",
-  },
-  download: {
-    title: "下載電腦版啟動器",
-    desc: "獲取獨立啟動器，以獲得更優化的效能和全螢幕遊戲體驗。",
-  },
-  comingSoon: {
-    placeholder: "敬請期待...",
-    featureComingSoon: "該功能即將上線，敬請期待。",
-  },
-  mode: {
-    onlineTitle: "線上遊玩",
-    onlineDesc: "登入伺服器。您的角色數據保存在伺服器上，與其他線上玩家共享同一個世界。",
-    onlineAria: "線上遊玩：登入持久共享的聯機世界",
-    offlineTitle: "單機試玩",
-    offlineDesc: "在瀏覽器中直接開啟單人世界。所有數據均不保存：非常適合快速試玩或功能測試。",
-    offlineAria: "單機試玩：直接開始本地單人遊戲會話",
-    tipTitle: "提示：",
-    tipText: "為了獲得最流暢的遊玩體驗，請在此網站上關閉廣告攔截外掛。根據社群回饋，部分攔截器可能會導致遊戲延遲。",
-  },
-  auth: {
-    enterRealm: "進入伺服器",
-    username: "使用者名稱",
-    usernameError: "請輸入使用者名稱。",
-    usernamePlaceholder: "請輸入使用者名稱",
-    password: "密碼",
-    passwordError: "請輸入密碼。",
-    passwordPlaceholder: "請輸入密碼",
-    showPassword: "顯示密碼",
-    hidePassword: "隱藏密碼",
-    logIn: "登入",
-    createAccount: "建立帳號",
-    back: "返回",
-    realmList: "伺服器列表",
-    loadingRealms: "正在載入伺服器...",
-    changeRealm: "切換伺服器",
-    characters: "角色列表：",
-    createCharacter: "建立角色",
-    characterName: "角色名稱",
-    characterNamePlaceholder: "輸入角色名稱",
-    enterWorld: "進入世界",
-    offlineCharacter: "單機角色",
-    create: "建立",
-  },
-  classes: {
-    warrior: "戰士",
-    paladin: "聖騎士",
-    hunter: "獵人",
-    rogue: "潛行者",
-    priest: "牧師",
-    shaman: "薩滿",
-    mage: "法師",
-    warlock: "術士",
-    druid: "德魯伊",
-    warriorAria: "戰士職業",
-    paladinAria: "聖騎士職業",
-    hunterAria: "獵人職業",
-    rogueAria: "潛行者職業",
-    priestAria: "牧師職業",
-    shamanAria: "薩滿職業",
-    mageAria: "法師職業",
-    warlockAria: "術士職業",
-    druidAria: "德魯伊職業",
-  },
-  controls: {
-    title: "操作指南",
-    movement: "移動操作",
-    moveTurn: "移動/轉向",
-    strafe: "向左/向右平移",
-    jump: "跳躍",
-    autorun: "切換自動奔跑",
-    combat: "戰鬥與互動",
-    target: "鎖定敵對目標",
-    spells: "施放法術",
-    interact: "互動/拾取",
-    nameplates: "顯示姓名板",
-    camera: "視角與滑鼠",
-    rightDrag: "右鍵拖曳",
-    leftDrag: "左鍵拖曳",
-    mouseWheel: "滑鼠滾輪",
-    mouselook: "滑鼠視角",
-    orbit: "環繞鏡頭",
-    zoom: "縮放",
-    interfaces: "介面視窗",
-    charPane: "角色面板",
-    spellbook: "法術書",
-    questLog: "任務日誌",
-    worldMap: "世界地圖",
-    bags: "背包背包欄",
-    emoteWheel: "按住表情輪盤",
-    friends: "好友與公會",
-    chat: "開啟對話框",
-  },
-  ...shellStrings.zh_TW,
-  ...hudStrings.zh_TW,
-  ...abilityStrings.zh_TW,
-  ...questStrings.zh_TW,
-  ...itemStrings.zh_TW,
-  ...classAbilityNames.zh_TW,
-  ...itemNames.zh_TW,
-  ...worldNames.zh_TW,
-  ...mergeStrings.zh_TW,
-  entities: {
-    ...itemNames.zh_TW.entities,
-    ...worldNames.zh_TW.entities,
-    abilities: { ...itemNames.zh_TW.entities.abilities, ...mergeExtra.zh_TW.abilities },
-    items: { ...itemNames.zh_TW.entities.items, ...mergeEntities.zh_TW.items, ...mergeExtra.zh_TW.items },
-    mobs: { ...worldNames.zh_TW.entities.mobs, ...mergeEntities.zh_TW.mobs, ...mergeExtra.zh_TW.mobs },
-    npcs: { ...worldNames.zh_TW.entities.npcs, ...mergeExtra.zh_TW.npcs },
-    quests: { ...worldNames.zh_TW.entities.quests, ...mergeEntities.zh_TW.quests, ...mergeExtra.zh_TW.quests },
-    dungeons: { ...worldNames.zh_TW.entities.dungeons, ...mergeExtra.zh_TW.dungeons },
-  },
-};
-
-export const ko_KR: typeof en = {
-  meta: { builtOn: "{date} 빌드" },
-  realmTypes: { normal: "일반", pvp: "PvP", rp: "롤플레잉", rpPvp: "롤플레잉 PvP" },
-  game: gameStringsKoKR,
-  nav: {
-    home: "홈",
-    play: "플레이",
-    stats: "통계",
-    about: "정보",
-    highscores: "순위표",
-    wiki: "위키",
-    news: "새소식",
-    download: "다운로드",
-    loginRegister: "로그인/회원가입",
-    donate: "후원",
-  },
-  stats: {
-    title: "렐름 상태",
-    accountsCreated: "생성된 계정 수",
-    playersOnline: "현재 접속자 수",
-    realmName: "렐름 이름",
-  },
-  footer: {
-    copyright: "2026 World of ClaudeCraft",
-    githubLink: "https://github.com/levy-street/world-of-claudecraft",
-    githubLabel: "오픈 소스 프로젝트",
-    terms: "서비스 이용약관",
-    privacy: "개인정보 처리방침",
-    discordLabel: "Discord 참여하기",
-  },
-  highscores: {
-    title: "최고 점수 순위표",
-    desc: "렐름의 가장 위대한 챔피언들을 확인하고 당신의 진행 상태를 비교해 보세요.",
-  },
-  wiki: {
-    title: "게임 위키 및 가이드",
-    desc: "렐름의 비밀과 클래스 가이드, 전략 등을 확인해 보세요.",
-  },
-  news: {
-    title: "새소식 및 업데이트",
-    desc: "최신 패치 노트, 이벤트 소식, 커뮤니티 업데이트를 확인해 보세요.",
-  },
-  download: {
-    title: "데스크톱 런처 다운로드",
-    desc: "최적화된 성능과 전체 화면 플레이를 위해 독립형 런처를 다운로드하세요.",
-  },
-  comingSoon: {
-    placeholder: "준비 중...",
-    featureComingSoon: "이 기능은 곧 렐름에 업데이트될 예정입니다.",
-  },
-  mode: {
-    onlineTitle: "온라인 플레이",
-    onlineDesc: "렐름에 로그인합니다. 생성한 캐릭터는 서버에 저장되며 온라인에 있는 다른 모든 플레이어와 같은 세상을 공유합니다.",
-    onlineAria: "온라인 플레이: 지속적으로 공유되는 렐름에 로그인합니다",
-    offlineTitle: "오프라인 플레이",
-    offlineDesc: "브라우저에서 즉시 일인용 세상을 플레이합니다. 아무것도 저장되지 않아 빠른 전투나 연습에 적합합니다.",
-    offlineAria: "오프라인 플레이: 즉각적인 로컬 싱글 플레이 세션을 시작합니다",
-    tipTitle: "팁:",
-    tipText: "가장 원활한 이용을 위해 이 웹사이트의 광고 차단기 확장 프로그램을 비활성화해 주세요. 일부 차단기가 게임 성능 지연을 유발할 수 있습니다.",
-  },
-  auth: {
-    enterRealm: "렐름 입장",
-    username: "사용자 이름",
-    usernameError: "사용자 이름을 입력해 주세요.",
-    usernamePlaceholder: "사용자 이름 입력",
-    password: "비밀번호",
-    passwordError: "비밀번호를 입력해 주세요.",
-    passwordPlaceholder: "비밀번호 입력",
-    showPassword: "비밀번호 표시",
-    hidePassword: "비밀번호 숨기기",
-    logIn: "로그인",
-    createAccount: "계정 생성",
-    back: "뒤로 가기",
-    realmList: "렐름 목록",
-    loadingRealms: "렐름을 불러오는 중...",
-    changeRealm: "렐름 변경",
-    characters: "캐릭터 목록:",
-    createCharacter: "캐릭터 생성",
-    characterName: "캐릭터 이름",
-    characterNamePlaceholder: "캐릭터 이름",
-    enterWorld: "세계 입장",
-    offlineCharacter: "오프라인 캐릭터",
-    create: "생성",
-  },
-  classes: {
-    warrior: "전사",
-    paladin: "성기사",
-    hunter: "사냥꾼",
-    rogue: "도적",
-    priest: "사제",
-    shaman: "주술사",
-    mage: "마법사",
-    warlock: "흑마법사",
-    druid: "드루이드",
-    warriorAria: "전사 직업",
-    paladinAria: "성기사 직업",
-    hunterAria: "사냥꾼 직업",
-    rogueAria: "도적 직업",
-    priestAria: "사제 직업",
-    shamanAria: "주술사 직업",
-    mageAria: "마법사 직업",
-    warlockAria: "흑마법사 직업",
-    druidAria: "드루이드 직업",
-  },
-  controls: {
-    title: "조작 가이드",
-    movement: "이동 조작",
-    moveTurn: "이동 / 회전",
-    strafe: "좌/우 평행 이동",
-    jump: "점프",
-    autorun: "자동 달리기 전환",
-    combat: "전투 및 상호작용",
-    target: "적 대상 지정",
-    spells: "주문 시전",
-    interact: "상호작용 / 전리품 획득",
-    nameplates: "이름표 전환",
-    camera: "카메라 및 마우스",
-    rightDrag: "우클릭 드래그",
-    leftDrag: "좌클릭 드래그",
-    mouseWheel: "마우스 휠",
-    mouselook: "마우스 시점",
-    orbit: "카메라 회전",
-    zoom: "줌",
-    interfaces: "인터페이스",
-    charPane: "캐릭터 정보",
-    spellbook: "마법책",
-    questLog: "퀘스트 목록",
-    worldMap: "세계 지도",
-    bags: "가방 소지품",
-    emoteWheel: "감정표현 휠 길게 누르기",
-    friends: "친구 및 길드",
-    chat: "채팅창 열기",
-  },
-  ...shellStrings.ko_KR,
-  ...hudStrings.ko_KR,
-  ...abilityStrings.ko_KR,
-  ...questStrings.ko_KR,
-  ...itemStrings.ko_KR,
-  ...classAbilityNames.ko_KR,
-  ...itemNames.ko_KR,
-  ...worldNames.ko_KR,
-  ...mergeStrings.ko_KR,
-  entities: {
-    ...itemNames.ko_KR.entities,
-    ...worldNames.ko_KR.entities,
-    abilities: { ...itemNames.ko_KR.entities.abilities, ...mergeExtra.ko_KR.abilities },
-    items: { ...itemNames.ko_KR.entities.items, ...mergeEntities.ko_KR.items, ...mergeExtra.ko_KR.items },
-    mobs: { ...worldNames.ko_KR.entities.mobs, ...mergeEntities.ko_KR.mobs, ...mergeExtra.ko_KR.mobs },
-    npcs: { ...worldNames.ko_KR.entities.npcs, ...mergeExtra.ko_KR.npcs },
-    quests: { ...worldNames.ko_KR.entities.quests, ...mergeEntities.ko_KR.quests, ...mergeExtra.ko_KR.quests },
-    dungeons: { ...worldNames.ko_KR.entities.dungeons, ...mergeExtra.ko_KR.dungeons },
-  },
-};
-
-export const ja_JP: typeof en = {
-  meta: { builtOn: "ビルド日：{date}" },
-  realmTypes: { normal: "ノーマル", pvp: "PvP", rp: "RP", rpPvp: "RP-PvP" },
-  game: gameStringsJaJP,
-  nav: {
-    home: "ホーム",
-    play: "プレイ",
-    stats: "統計",
-    about: "情報",
-    highscores: "ハイスコア",
-    wiki: "Wiki",
-    news: "ニュース",
-    download: "ダウンロード",
-    loginRegister: "ログイン/登録",
-    donate: "寄付",
-  },
-  stats: {
-    title: "レルムステータス",
-    accountsCreated: "作成アカウント数",
-    playersOnline: "オンラインプレイヤー数",
-    realmName: "レルム名",
-  },
-  footer: {
-    copyright: "2026 World of ClaudeCraft",
-    githubLink: "https://github.com/levy-street/world-of-claudecraft",
-    githubLabel: "オープンソースプロジェクト",
-    terms: "利用規約",
-    privacy: "プライバシーポリシー",
-    discordLabel: "Discordに参加する",
-  },
-  highscores: {
-    title: "ハイスコアリーダーボード",
-    desc: "レルムの偉大なチャンピオンたちを確認し、あなたの進行度と比較しましょう。",
-  },
-  wiki: {
-    title: "ゲームWiki & ガイド",
-    desc: "レルムの秘密、クラスガイド、戦略を確認しましょう。",
-  },
-  news: {
-    title: "ニュース & アップデート",
-    desc: "最新のパッチノート、イベント、コミュニティの更新情報を確認しましょう。",
-  },
-  download: {
-    title: "デスクトップランチャーのダウンロード",
-    desc: "最適化されたパフォーマンスと全画面表示でのプレイのために、スタンドアロン版ランチャーを入手してください。",
-  },
-  comingSoon: {
-    placeholder: "近日公開...",
-    featureComingSoon: "この機能はまもなくレルムに実装されます。",
-  },
-  mode: {
-    onlineTitle: "オンラインでプレイ",
-    onlineDesc: "レルムにログインします。キャラクターはサーバー上に保存され、オンラインの他の全プレイヤーと同じ世界を共有します。",
-    onlineAria: "オンラインでプレイ：永続的な共有レルムにログインします",
-    offlineTitle: "オフラインでプレイ",
-    offlineDesc: "ブラウザ上で即座にシングルプレイヤーの世界を起動します。保存は一切されず、小規模な対戦やテストに最適です。",
-    offlineAria: "オフラインでプレイ：即座にローカルシングルプレイヤーセッションを開始します",
-    tipTitle: "ヒント：",
-    tipText: "よりスムーズな体験のために、本サイトでの広告ブロック拡張機能を無効にしてください。コミュニティの報告によると、一部のブロック機能がラグの原因になる場合があります。",
-  },
-  auth: {
-    enterRealm: "レルムに入る",
-    username: "ユーザー名",
-    usernameError: "ユーザー名を入力してください。",
-    usernamePlaceholder: "ユーザー名を入力",
-    password: "パスワード",
-    passwordError: "パスワードを入力してください。",
-    passwordPlaceholder: "パスワードを入力",
-    showPassword: "パスワードを表示",
-    hidePassword: "パスワードを非表示",
-    logIn: "ログイン",
-    createAccount: "アカウント作成",
-    back: "戻る",
-    realmList: "レルムリスト",
-    loadingRealms: "レルムを読み込み中...",
-    changeRealm: "レルム変更",
-    characters: "キャラクター数：",
-    createCharacter: "キャラクター作成",
-    characterName: "キャラクター名",
-    characterNamePlaceholder: "キャラクター名",
-    enterWorld: "世界に入る",
-    offlineCharacter: "オフラインキャラクター",
-    create: "作成",
-  },
-  classes: {
-    warrior: "戦士",
-    paladin: "パラディン",
-    hunter: "ハンター",
-    rogue: "ローグ",
-    priest: "プリースト",
-    shaman: "シャーマン",
-    mage: "メイジ",
-    warlock: "ウォーロック",
-    druid: "ドルイド",
-    warriorAria: "戦士クラス",
-    paladinAria: "パラディンクラス",
-    hunterAria: "ハンタークラス",
-    rogueAria: "ローグクラス",
-    priestAria: "プリーストクラス",
-    shamanAria: "シャーマンクラス",
-    mageAria: "メイジクラス",
-    warlockAria: "ウォーロッククラス",
-    druidAria: "ドルイドクラス",
-  },
-  controls: {
-    title: "操作ガイド",
-    movement: "移動",
-    moveTurn: "移動 / 旋回",
-    strafe: "左右平行移動",
-    jump: "ジャンプ",
-    autorun: "オートラン切り替え",
-    combat: "戦闘 & 相互作用",
-    target: "敵をターゲット",
-    spells: "呪文の詠唱",
-    interact: "相互作用 / 戦利品回収",
-    nameplates: "ネームプレート表示切り替え",
-    camera: "カメラ & マウス",
-    rightDrag: "右ドラッグ",
-    leftDrag: "左ドラッグ",
-    mouseWheel: "マウスホイール",
-    mouselook: "マウスルック",
-    orbit: "カメラの旋回",
-    zoom: "ズーム",
-    interfaces: "画面表示",
-    charPane: "キャラクター画面",
-    spellbook: "魔道書",
-    questLog: "クエストログ",
-    worldMap: "ワールドマップ",
-    bags: "バッグのインベントリ",
-    emoteWheel: "エモートホイールを長押し",
-    friends: "フレンド & ギルド",
-    chat: "チャットを開く",
-  },
-  ...shellStrings.ja_JP,
-  ...hudStrings.ja_JP,
-  ...abilityStrings.ja_JP,
-  ...questStrings.ja_JP,
-  ...itemStrings.ja_JP,
-  ...classAbilityNames.ja_JP,
-  ...itemNames.ja_JP,
-  ...worldNames.ja_JP,
-  ...mergeStrings.ja_JP,
-  entities: {
-    ...itemNames.ja_JP.entities,
-    ...worldNames.ja_JP.entities,
-    abilities: { ...itemNames.ja_JP.entities.abilities, ...mergeExtra.ja_JP.abilities },
-    items: { ...itemNames.ja_JP.entities.items, ...mergeEntities.ja_JP.items, ...mergeExtra.ja_JP.items },
-    mobs: { ...worldNames.ja_JP.entities.mobs, ...mergeEntities.ja_JP.mobs, ...mergeExtra.ja_JP.mobs },
-    npcs: { ...worldNames.ja_JP.entities.npcs, ...mergeExtra.ja_JP.npcs },
-    quests: { ...worldNames.ja_JP.entities.quests, ...mergeEntities.ja_JP.quests, ...mergeExtra.ja_JP.quests },
-    dungeons: { ...worldNames.ja_JP.entities.dungeons, ...mergeExtra.ja_JP.dungeons },
-  },
-};
-
-export const pt_BR: typeof en = {
-  meta: { builtOn: "Compilado em {date}" },
-  realmTypes: { normal: "Normal", pvp: "PvP", rp: "RP", rpPvp: "RP-PvP" },
-  game: gameStringsPtBR,
-  nav: {
-    home: "Início",
-    play: "Jogar",
-    stats: "Estatísticas",
-    about: "Sobre",
-    highscores: "Recordes",
-    wiki: "Wiki",
-    news: "Notícias",
-    download: "Baixar",
-    loginRegister: "Entrar/Registrar",
-    donate: "Doar",
-  },
-  stats: {
-    title: "Status do Reino",
-    accountsCreated: "Contas criadas",
-    playersOnline: "Jogadores online",
-    realmName: "Nome do reino",
-  },
-  footer: {
-    copyright: "2026 World of ClaudeCraft",
-    githubLink: "https://github.com/levy-street/world-of-claudecraft",
-    githubLabel: "Projeto de código aberto",
-    terms: "Termos de serviço",
-    privacy: "Política de privacidade",
-    discordLabel: "Juntar-se ao Discord",
-  },
-  highscores: {
-    title: "Tabela de recordes",
-    desc: "Acompanhe os maiores campeões do reino e compare o seu progresso.",
-  },
-  wiki: {
-    title: "Wiki e guia do jogo",
-    desc: "Descubra os segredos do reino, guias de classes e estratégias.",
-  },
-  news: {
-    title: "Notícias e atualizações",
-    desc: "Leia as últimas notas do patch, eventos e atualizações da comunidade.",
-  },
-  download: {
-    title: "Baixar launcher para desktop",
-    desc: "Obtenha o launcher independente para desempenho otimizado e jogo em tela cheia.",
-  },
-  comingSoon: {
-    placeholder: "Em breve...",
-    featureComingSoon: "Este recurso estará disponível no reino em breve.",
-  },
-  mode: {
-    onlineTitle: "Jogar online",
-    onlineDesc: "Entre no reino. Seus personagens vivem no servidor e você compartilha o mundo com todos os outros online.",
-    onlineAria: "Jogar online: entre no reino compartilhado persistente",
-    offlineTitle: "Jogar offline",
-    offlineDesc: "Mundo solo instantâneo no seu navegador. Nada é salvo: perfeito para uma luta rápida ou testes.",
-    offlineAria: "Jogar offline: inicie uma sessão local de um jogador instantânea",
-    tipTitle: "DICA:",
-    tipText: "Para uma experiência mais suave, desative os bloqueadores de anúncios neste site. Relatos da comunidade indicam que alguns bloqueadores podem causar lag.",
-  },
-  auth: {
-    enterRealm: "Entrar no reino",
-    username: "Nome de usuário",
-    usernameError: "Por favor, insira seu nome de usuário.",
-    usernamePlaceholder: "Insira o usuário",
-    password: "Senha",
-    passwordError: "Por favor, insira sua senha.",
-    passwordPlaceholder: "Insira a senha",
-    showPassword: "Mostrar senha",
-    hidePassword: "Ocultar senha",
-    logIn: "Entrar",
-    createAccount: "Criar conta",
-    back: "Voltar",
-    realmList: "Lista de reinos",
-    loadingRealms: "Carregando reinos...",
-    changeRealm: "Mudar de reino",
-    characters: "Personagens:",
-    createCharacter: "Criar personagem",
-    characterName: "Nome do personagem",
-    characterNamePlaceholder: "Nome do personagem",
-    enterWorld: "Entrar no mundo",
-    offlineCharacter: "Personagem offline",
-    create: "Criar",
-  },
-  classes: {
-    warrior: "Guerreiro",
-    paladin: "Paladino",
-    hunter: "Caçador",
-    rogue: "Ladino",
-    priest: "Sacerdote",
-    shaman: "Xamã",
-    mage: "Mago",
-    warlock: "Bruxo",
-    druid: "Druida",
-    warriorAria: "Classe Guerreiro",
-    paladinAria: "Classe Paladino",
-    hunterAria: "Classe Caçador",
-    rogueAria: "Classe Ladino",
-    priestAria: "Classe Sacerdote",
-    shamanAria: "Classe Xamã",
-    mageAria: "Classe Mago",
-    warlockAria: "Classe Bruxo",
-    druidAria: "Classe Druida",
-  },
-  controls: {
-    title: "Guia de controles",
-    movement: "Movimentação",
-    moveTurn: "Mover / Virar",
-    strafe: "Passo lateral Esquerda/Direita",
-    jump: "Pular",
-    autorun: "Alternar corrida automática",
-    combat: "Combate e interação",
-    target: "Selecionar inimigo como alvo",
-    spells: "Conjurar feitiços",
-    interact: "Interagir / Saquear",
-    nameplates: "Alternar placas de nome",
-    camera: "Câmera e mouse",
-    rightDrag: "Arrastar com botão direito",
-    leftDrag: "Arrastar com botão esquerdo",
-    mouseWheel: "Roda do mouse",
-    mouselook: "Olhar com o mouse",
-    orbit: "Orbitar câmera",
-    zoom: "Zoom",
-    interfaces: "Interfaces",
-    charPane: "Painel do personagem",
-    spellbook: "Livro de feitiços",
-    questLog: "Registro de missões",
-    worldMap: "Mapa do mundo",
-    bags: "Inventário de bolsas",
-    emoteWheel: "Segurar roda de emotes",
-    friends: "Amigos e guilda",
-    chat: "Abrir chat",
-  },
-  ...shellStrings.pt_BR,
-  ...hudStrings.pt_BR,
-  ...abilityStrings.pt_BR,
-  ...questStrings.pt_BR,
-  ...itemStrings.pt_BR,
-  ...classAbilityNames.pt_BR,
-  ...itemNames.pt_BR,
-  ...worldNames.pt_BR,
-  ...mergeStrings.pt_BR,
-  entities: {
-    ...itemNames.pt_BR.entities,
-    ...worldNames.pt_BR.entities,
-    abilities: { ...itemNames.pt_BR.entities.abilities, ...mergeExtra.pt_BR.abilities },
-    items: { ...itemNames.pt_BR.entities.items, ...mergeEntities.pt_BR.items, ...mergeExtra.pt_BR.items },
-    mobs: { ...worldNames.pt_BR.entities.mobs, ...mergeEntities.pt_BR.mobs, ...mergeExtra.pt_BR.mobs },
-    npcs: { ...worldNames.pt_BR.entities.npcs, ...mergeExtra.pt_BR.npcs },
-    quests: { ...worldNames.pt_BR.entities.quests, ...mergeEntities.pt_BR.quests, ...mergeExtra.pt_BR.quests },
-    dungeons: { ...worldNames.pt_BR.entities.dungeons, ...mergeExtra.pt_BR.dungeons },
-  },
-};
-
-export const ru_RU: typeof en = {
-  meta: { builtOn: "Сборка от {date}" },
-  realmTypes: { normal: "Обычный", pvp: "PvP", rp: "Ролевой", rpPvp: "Ролевой-PvP" },
-  game: gameStringsRuRU,
-  nav: {
-    home: "Главная",
-    play: "Играть",
-    stats: "Статистика",
-    about: "О проекте",
-    highscores: "Таблицы рекордов",
-    wiki: "Вики",
-    news: "Новости",
-    download: "Скачать",
-    loginRegister: "Войти / Зарегистрироваться",
-    donate: "Пожертвовать",
-  },
-  stats: {
-    title: "Статус игрового мира",
-    accountsCreated: "Создано учетных записей",
-    playersOnline: "Игроков в сети",
-    realmName: "Название игрового мира",
-  },
-  footer: {
-    copyright: "2026 World of ClaudeCraft",
-    githubLink: "https://github.com/levy-street/world-of-claudecraft",
-    githubLabel: "Проект с открытым исходным кодом",
-    terms: "Условия использования",
-    privacy: "Политика конфиденциальности",
-    discordLabel: "Присоединиться к Discord",
-  },
-  highscores: {
-    title: "Таблица рекордов",
-    desc: "Отслеживайте величайших чемпионов игрового мира и сравнивайте свои успехи.",
-  },
-  wiki: {
-    title: "Вики и руководство",
-    desc: "Откройте секреты игрового мира, руководства по классам и стратегии.",
-  },
-  news: {
-    title: "Новости и обновления",
-    desc: "Читайте описание последних обновлений, информацию о событиях и новости сообщества.",
-  },
-  download: {
-    title: "Скачать загрузчик для ПК",
-    desc: "Загрузите отдельный клиент для оптимальной производительности и игры на весь экран.",
-  },
-  comingSoon: {
-    placeholder: "Скоро...",
-    featureComingSoon: "Эта функция скоро появится в игре.",
-  },
-  mode: {
-    onlineTitle: "Играть по сети",
-    onlineDesc: "Войдите в игровой мир. Ваши персонажи сохраняются на сервере, а вы делите игровой мир с другими игроками в сети.",
-    onlineAria: "Играть по сети: вход в постоянный общий игровой мир",
-    offlineTitle: "Играть вне сети",
-    offlineDesc: "Мгновенный одиночный мир прямо в браузере. Прогресс не сохраняется: идеально для быстрой драки или тестирования.",
-    offlineAria: "Играть вне сети: запуск мгновенного локального одиночного сеанса",
-    tipTitle: "СОВЕТ:",
-    tipText: "Для максимально плавной игры отключите блокировщики рекламы на этом сайте. По отзывам сообщества, некоторые блокировщики могут вызывать задержки.",
-  },
-  auth: {
-    enterRealm: "Войти в игровой мир",
-    username: "Имя пользователя",
-    usernameError: "Пожалуйста, введите имя пользователя.",
-    usernamePlaceholder: "Введите имя пользователя",
-    password: "Пароль",
-    passwordError: "Пожалуйста, введите пароль.",
-    passwordPlaceholder: "Введите пароль",
-    showPassword: "Показать пароль",
-    hidePassword: "Скрыть пароль",
-    logIn: "Войти",
-    createAccount: "Создать учетную запись",
-    back: "Назад",
-    realmList: "Список игровых миров",
-    loadingRealms: "Загрузка списка миров...",
-    changeRealm: "Сменить игровой мир",
-    characters: "Персонажи:",
-    createCharacter: "Создать персонажа",
-    characterName: "Имя персонажа",
-    characterNamePlaceholder: "Имя персонажа",
-    enterWorld: "Войти в мир",
-    offlineCharacter: "Персонаж вне сети",
-    create: "Создать",
-  },
-  classes: {
-    warrior: "Воин",
-    paladin: "Паладин",
-    hunter: "Охотник",
-    rogue: "Разбойник",
-    priest: "Жрец",
-    shaman: "Шаман",
-    mage: "Маг",
-    warlock: "Чернокнижник",
-    druid: "Друид",
-    warriorAria: "Класс Воин",
-    paladinAria: "Класс Паладин",
-    hunterAria: "Класс Охотник",
-    rogueAria: "Класс Разбойник",
-    priestAria: "Класс Жрец",
-    shamanAria: "Класс Шаман",
-    mageAria: "Класс Маг",
-    warlockAria: "Класс Чернокнижник",
-    druidAria: "Класс Друид",
-  },
-  controls: {
-    title: "Руководство по управлению",
-    movement: "Передвижение",
-    moveTurn: "Движение / поворот",
-    strafe: "Стрейф влево/вправо",
-    jump: "Прыжок",
-    autorun: "Вкл/выкл автобег",
-    combat: "Бой и взаимодействие",
-    target: "Выбор врага",
-    spells: "Заклинания",
-    interact: "Взаимодействие / сбор добычи",
-    nameplates: "Вкл/выкл индикаторы здоровья",
-    camera: "Камера и мышь",
-    rightDrag: "Перетаскивание правой кнопкой",
-    leftDrag: "Перетаскивание левой кнопкой",
-    mouseWheel: "Колесико мыши",
-    mouselook: "Обзор мышью",
-    orbit: "Вращение камеры",
-    zoom: "Приближение",
-    interfaces: "Интерфейсы",
-    charPane: "Окно персонажа",
-    spellbook: "Книга заклинаний",
-    questLog: "Журнал заданий",
-    worldMap: "Карта мира",
-    bags: "Сумки",
-    emoteWheel: "Удерживать колесо эмоций",
-    friends: "Друзья и гильдия",
-    chat: "Открыть чат",
-  },
-  ...shellStrings.ru_RU,
-  ...hudStrings.ru_RU,
-  ...abilityStrings.ru_RU,
-  ...questStrings.ru_RU,
-  ...itemStrings.ru_RU,
-  ...classAbilityNames.ru_RU,
-  ...itemNames.ru_RU,
-  ...worldNames.ru_RU,
-  ...mergeStrings.ru_RU,
-  entities: {
-    ...itemNames.ru_RU.entities,
-    ...worldNames.ru_RU.entities,
-    abilities: { ...itemNames.ru_RU.entities.abilities, ...mergeExtra.ru_RU.abilities },
-    items: { ...itemNames.ru_RU.entities.items, ...mergeEntities.ru_RU.items, ...mergeExtra.ru_RU.items },
-    mobs: { ...worldNames.ru_RU.entities.mobs, ...mergeEntities.ru_RU.mobs, ...mergeExtra.ru_RU.mobs },
-    npcs: { ...worldNames.ru_RU.entities.npcs, ...mergeExtra.ru_RU.npcs },
-    quests: { ...worldNames.ru_RU.entities.quests, ...mergeEntities.ru_RU.quests, ...mergeExtra.ru_RU.quests },
-    dungeons: { ...worldNames.ru_RU.entities.dungeons, ...mergeExtra.ru_RU.dungeons },
-  },
-};
 
 const translations = {
   en,
